### 2022-02-08 #4138 @whedon
[PRE REVIEW]: CVtreeMLE: Efficient Estimation of Mixed Exposures using Data Adaptive Decision Trees and Cross-Validated Targeted Maximum Likelihood Estimation in R
**Submitting author:** @blind-contours (<a href="http://orcid.org/0000-0002-5515-6307">David McCoy</a>)
**Repository:** <a href="https://github.com/blind-contours/CVtreeMLE" target ="_blank">https://github.com/blind-contours/CVtreeMLE</a>
**Version:** v1.0.0
