### 2022-02-11 #4138 @KristinaRiemer
@kthyng this is pretty far outside my statistical background, so I think for that reason alone it might be better to find a more suitable editor. I'm also a little swamped right now and have been meaning to modify my editorial load accordingly.
