### 2022-02-12 #4138 @osorensen
@blind-contours, I am the associate editor for this submission now, and will try to find reviewers. In the meantime, could you please correct the following in the submitted manuscript:

- Line 63: rewrite "represenation" to "representation".
- Lines 77-78: I assume "general additive models" should be "generalized additive models"?
