### 2022-02-14 27a77c8
paper edits spotted by JOSS
