+++ paper/paper.md
@@ -38,11 +38,11 @@ In many disciplines there is a demonstrable need to ascertain the causal effects
 
 # Background
 
-In most research scenarios, the analyst is interested in causal inference for an **a priori** specified treatment or exposure. However, in the evaluation of a mixed exposure, such as air pollution or pesticides, it is not possible to estimate the expected outcome given every combination of exposures due to high-dimensionality and sparsity. Even if this approach were possible, a target parameter that can inform public policy is still ill-defined. In such a setting, it is helpful to map a set of continuous mixture components into a lower dimensional represenation of exposure using a pre-determined algorithm to estimate a target parameter that has a meaningful interpretation. Decision trees provide a useful solution by mapping a set of exposures into a rule which can be represented as a binary vector. This binary vector indicates whether an individual has been exposed to a particular rule estimated by the decision tree. Our target parameter is then defined as the mean difference in counterfactual outcomes for those exposed to the mixture rule compared to those unexposed, or the average treatment effect for the mixture.
+In most research scenarios, the analyst is interested in causal inference for an **a priori** specified treatment or exposure. However, in the evaluation of a mixed exposure, such as air pollution or pesticides, it is not possible to estimate the expected outcome given every combination of exposures due to high-dimensionality and sparsity. Even if this approach were possible, a target parameter that can inform public policy is still ill-defined. In such a setting, it is helpful to map a set of continuous mixture components into a lower dimensional representation of exposure using a pre-determined algorithm to estimate a target parameter that has a meaningful interpretation. Decision trees provide a useful solution by mapping a set of exposures into a rule which can be represented as a binary vector. This binary vector indicates whether an individual has been exposed to a particular rule estimated by the decision tree. Our target parameter is then defined as the mean difference in counterfactual outcomes for those exposed to the mixture rule compared to those unexposed, or the average treatment effect for the mixture.
 
 # `CVtreeMLE`'s Scope
 
-Building on prior work related to data-adaptive parameters [@Hubbard2016] and CV-TMLE [@Zheng2010], `CVtreeMLE` is a novel approach for estimating the joint impact of a mixed exposure by using cross-validated targeted minimum loss-based estimation which guarantees consistency, efficiency, and multiple robustness despite using highly flexible learners to estimate a data-adaptive parameter. `CVtreeMLE` summarizes the effect of a joint exposure on the outcome of interest by first doing an iterative backfitting procedure, similar to general additive models, to fit f(A), a Super Learner of decision trees, and h(W), an unrestricted Super Learner, in a semi-parametric model; E(Y|A,W) = f(A) + h(W), where A is a vector of exposures and W is a vector of covariates. In this way, we can data-adaptively find the best fitting decision tree model which has the lowest cross-validated model error while flexibly adjusting for covariates. This procedure is done to find rules for the mixture modeled collectively and for each mixture component individually. There are two types of results, 1. an average treatment effect (ATE) comparing those who fall within a subspace of the joint exposure versus those in the complement of that space and 2. the ATE for each data-adaptively identified threshold of an individual mixture component when compared to the lowest identified exposure level. The `CVtreeMLE` software package, for the `R` language and environment for statistical computing [@R], implements this methodology for deriving causal inference from ensemble decision trees.
[…]
