### 2022-02-14 #4138 @blind-contours
> @blind-contours, I am the associate editor for this submission now, and will try to find reviewers. In the meantime, could you please correct the following in the submitted manuscript:
>
> * Line 63: rewrite "represenation" to "representation".
> * Lines 77-78: I assume "general additive models" should be "generalized additive models"?

@osorensen thanks for spotting these - I have made the corrections.
