### 2022-02-18 #4181 @whedon
[REVIEW]: CVtreeMLE: Efficient Estimation of Mixed Exposures using Data Adaptive Decision Trees and Cross-Validated Targeted Maximum Likelihood Estimation in R
**Submitting author:** @blind-contours (<a href="http://orcid.org/0000-0002-5515-6307">David McCoy</a>)
**Repository:** https://github.com/blind-contours/CVtreeMLE
**Branch with paper.md** (empty if default branch):
**Version:** v1.0-joss
