### 2022-02-23 fa2638c
Updates based on windows error and no longer filtering based on only consistent rules.
