### 2022-02-23 f00b6f5
slurm script
