### 2022-02-23 15f7043
slurm job fails in foreach no error
