### 2022-02-23 3f8647e
Removed quantile sum g-comp from namespace and vignette.
