### 2022-02-23 3109bb1
Update R-CMD-check.yml
bayesian learner arm package added to yml
