+++ R/CVtreeMLE.R
@@ -13,6 +13,7 @@
 #' @param max_iter Max number of iterations of iterative backfitting algorithm
 #' @param verbose If true, creates a sink file where detailed information and diagnostics of the run process are given.
 #' @param parallel Use parallel processing if a backend is registered; enabled by default.
+#' @param parallel_cv Use parallel processing on CV procedure vs. parallel processing on Super Learner model fitting
 #' @param num_cores If using parallel, the number of cores to parallelize over
 
 #' @return Results object. TODO: add more detail here.
@@ -78,6 +79,7 @@ CVtreeMLE <- function(W,
                       family,
                       H.AW_trunc_lvl = 10,
                       parallel = TRUE,
+                      parallel_cv = TRUE,
                       num_cores,
                       max_iter = 5,
                       verbose = FALSE) {
@@ -158,27 +160,55 @@ CVtreeMLE <- function(W,
 
   ###################### Fit Iterative Backwards Rule Ensembles to the Mixture Components Across Fold Data ########################################
 
-  fold_mixture_rules <- foreach::foreach(fold_k = unique(data$folds), .combine = "rbind", .multicombine = TRUE) %dopar% {
-    ## get training and validation samples
+  if (parallel_cv == TRUE) {
 
-    At <- data[data$folds != fold_k, ]
-    Av <- data[data$folds == fold_k, ]
+    fold_mixture_rules <- foreach::foreach(fold_k = unique(data$folds), .combine = "rbind", .multicombine = TRUE) %dopar% {
+      ## get training and validation samples
 
-    rules <-
-      fit_iterative_mix_rule_backfitting(
-        At = At,
-        A = A,
-        W = W,
-        Y = Y,
-        Q1_stack = back_iter_SL,
-        fold = fold_k,
-        max_iter,
-        verbose
-      )
+      At <- data[data$folds != fold_k, ]
+      Av <- data[data$folds == fold_k, ]
 
-    rules
+      rules <-
+        fit_iterative_mix_rule_backfitting(
+          At = At,
+          A = A,
+          W = W,
+          Y = Y,
+          Q1_stack = back_iter_SL,
+          fold = fold_k,
+          max_iter,
+          verbose,
+          parallel_cv
+        )
+
+      rules
+    }
+
+  }else{
+
+    fold_mixture_rules <- foreach::foreach(fold_k = unique(data$folds), .combine = "rbind", .multicombine = TRUE) %do% {
+      ## get training and validation samples
+
+      At <- data[data$folds != fold_k, ]
+      Av <- data[data$folds == fold_k, ]
+
+      rules <-
+        fit_iterative_mix_rule_backfitting(
+          At = At,
+          A = A,
+          W = W,
+          Y = Y,
+          Q1_stack = back_iter_SL,
+          fold = fold_k,
+          max_iter,
+          verbose,
+          parallel_cv
+        )
+
+      rules
+    }
+
   }
-
   # remove the rows where no rules were found
   fold_mixture_rules <- fold_mixture_rules[fold_mixture_rules$description != "No Rules Found", ]
 
@@ -198,26 +228,53 @@ CVtreeMLE <- function(W,
 
   ###################### Fit Iterative Backwards Rule Ensembles to the Marginal Mixture Components Across Fold Data ########################################
 
-  fold_marginal_rules <- foreach::foreach(fold_k = unique(data$folds), .combine = "rbind", .multicombine = TRUE) %dopar% {
-    ## get training and validation samples
+  if (parallel_cv == TRUE) {
 
[…]
