+++ sandbox/03_run_simulation.R
@@ -1,23 +1,20 @@
-# be sure to set this env variable by `export R_LIBDIR=/path/to/your/R/libs`
-r_libdir <- Sys.getenv("R_LIBDIR")
-
-# set user-specific package library
-if (grepl("savio2", Sys.info()["nodename"])) {
-  .libPaths(r_libdir)
-  Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS="true")
-}
+# # be sure to set this env variable by `export R_LIBDIR=/path/to/your/R/libs`
+# r_libdir <- Sys.getenv("R_LIBDIR")
+#
+# # set user-specific package library
+# if (grepl("savio2", Sys.info()["nodename"])) {
+#   .libPaths(r_libdir)
+#   Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS="true")
+# }
 
 # packages
 library(here)
-library(foreach)
-library(future)
-library(doFuture)
-library(doRNG)
 library(data.table)
 library(tidyverse)
 library(origami)
 library(sl3)
 library(dplyr)
+
 
 devtools::load_all(here())
 
