### 2022-02-27 blind-contours/CVtreeMLE#2 @GaryBAYLOR
Why two licenses
I noticed there are two license files for the package. Can you explain besides MIT license, what's the license in the plain file `LICENSE`, and what's the purpose of having two licenses for the package?
