### 2022-02-27 blind-contours/CVtreeMLE#3 @GaryBAYLOR
The package cannot be found in CRAN
The package cannot be accessed via link https://CRAN.R-project.org/package=CVtreeMLE or https://cran.r-project.org/web/packages/CVtreeMLE/index.html, and installation by running `install.packages("CVtreeMLE")`, as instructed in README file, will fail.

Please check if the package has been successfully submitted to CRAN.
