### 2022-02-28 blind-contours/CVtreeMLE#4 @GaryBAYLOR
Improve function documentation
The documentation for the functions (`Rd` files) needs to be improved. The issues I can identify include

* `\title{}`: Be consistent in title capitalization by either capitalizing the first word or all words in the title. As an example, the current title for the function `bound_propensity` is *Bound Generalized Propensity Score*, while the title for the function `calc_additive_ate` is *Evaluate mixture rules found during the rpart decision tree process*

* `\description{}`: Do not simply repeat the title but provide a short description of what the function(s) do(es) (one paragraph, a few lines only)

* `\value{}`: Finish TODOs

* `\examples{}`: Add one or more examples to demo how to use each function. It is also used for the purpose of function testing purpose. Please refer to https://cran.r-project.org/doc/manuals/R-exts.html#Rd-format for more details
