### 2022-03-01 blind-contours/CVtreeMLE#5 @GaryBAYLOR
Lightweight editing of the JOSS paper
`line 10:` glm --> GLM
`line 30:` Thus --> Therefore
`line 39`: glm's --> GLMs
`line 48, line 50`: semiparametric --> semi-parametric
`line 58`: unbold "a priori" and change it to italic format
`line 69`: average treatment effect --> average treatment effect (ATE).  The reason is, when using an acronym, state it when it first appears in the paper and only use the acronym through the rest of the paper.
`line 84`: average treatment effect (ATE) --> ATE
`line 87-88`:  Remove 'language and environment for statistical computing'
`line 91`: CVtreeMLE --> It
`line 92`: sl3package --> sl3 package
`line 95-96`: the hyperlink to CRAN is invalid as the package hasn't been published there.
`line 102-103`: To make sure the title of the reference matches with the title in DOI, sl3: Modern pipelines for machine learning and Super Learning --> sl3: Modern super learning with pipelines
`line 105-106`: Statistical Inference for Data Adaptive Target Parameters --> Statistical inference for data adaptive target parameters.  This is to make all the references look consistent.
