### 2022-03-01 blind-contours/CVtreeMLE#8 @GaryBAYLOR
Comments on README file
* Simulation plots
In the plot that explains the simulation settings, please explain what does the plot on the left (with orange cube) represent (co-variables? then why three dimensions?), what is the relationship between A, W and M1, M2, M3? What is the difference between M1, M2,M3 in the left plot with those in the right plot?

* number of folds in cross-validation
It makes sense to use a higher value of `n_folds` instead of 2 to see how the results look like, as typically in cross-validation, higher number of folds(e.g., `n_folds` = 4, 8, 10, etc.) are used.
