### 2022-03-02 #4181 @GaryBAYLOR
@osorensen I reviewed the paper and have the following comments. I will take another look after these issues are resolved. Thanks!

Why two licenses https://github.com/blind-contours/CVtreeMLE/issues/2
The package cannot be found in CRAN https://github.com/blind-contours/CVtreeMLE/issues/3
Improve function documentation https://github.com/blind-contours/CVtreeMLE/issues/4
Lightweight editing of the JOSS paper https://github.com/blind-contours/CVtreeMLE/issues/5
Comments on the paper https://github.com/blind-contours/CVtreeMLE/issues/6
Remove R code that are no longer needed https://github.com/blind-contours/CVtreeMLE/issues/7
Comments on README file https://github.com/blind-contours/CVtreeMLE/issues/8
Non reproducible results https://github.com/blind-contours/CVtreeMLE/issues/9
