### 2022-03-12 #4181 @osorensen
Or perhaps you could try running the following comment in this issue, @cpalmer718: `@editorialbot generate my checklist`. This should generate a new checklist, which you can edit.
