### 2022-03-12 blind-contours/CVtreeMLE#10 @cpalmer718
Test coverage
the test coverage from `covr` is ~71% but it seems like the coverage is somewhat more accurately ~4%. `tests/testthat/test_CVtreeMLE_breastcancer.R` runs an example but doesn't include any assertions and doesn't inspect the output of the run (and in fact emits a warning on my system. due to how the test files are structured, the only function that seems to have substantial coverage right now is `simulate_mixture_cube`.

with the above, I find it very challenging to evaluate the functionality of the software. please expand the test coverage as necessary to ensure the majority of the functions are running not just end to end, but as expected when run.
