### 2022-03-12 blind-contours/CVtreeMLE#8 @ghost
contributions: the contribution link `https://github.com/blind-contours/CVtreeMLE/blob/master/CONTRIBUTING.md` does not lead anywhere
