### 2022-03-12 blind-contours/CVtreeMLE#4 @ghost
I concur about the above. I'll add that checking the `R CMD CHECK` output is helpful to spot some obvious ones, like missing parameter documentation. The github action unfortunately only errors on actual errors, whereas the warnings and notes can be very helpful.
