### 2022-03-12 blind-contours/CVtreeMLE#8 @ghost
I'll add here the following:

  - there's substantial example usage for simulation, both here and in the vignette, but a real world example would be really helpful
  - since there's such integration with SL3, a brief discussion of intended use (in text, not code) would probably be warranted, especially because it seems like some of the package's behavioral guarantees really rely on that package
