### 2022-03-12 blind-contours/CVtreeMLE#11 @cpalmer718
Installation of dev version
README lists `devel` branch for development version of package, but there is no such branch and as such the installation fails via that method.
