### 2022-03-12 blind-contours/CVtreeMLE#12 @ghost
Additional specific comments:

- the link to the personal webpage at the top leads to a webpage with generic placeholder content
- the vignette heavily cites van der Laan; this is perfectly reasonable, but at some point here or in the paper, additional citations from other sources would be appropriate to add, to help contextualize the work in the broader field
- `g(W)` appears in `Iterative Backfitting of Decision Trees` without definition, and is probably a typo. I'd recommend doing a pass over all the variable definitions and formulae to make sure they're clear and typo-free, and rendered correctly in markdown
- `h(A,W)` and `f(A,W)` appear without definition, after `h` and `f` had been defined marginally. I think I understand what you're going for with that, but given that, if it's what I think, it's really important for understanding, I think making sure really that whole section is clear is essential
- the document makes several assertions of asymptotic performance, but doesn't define what asymptotes are being mentioned. I think being as clear as possible with this in particular will help the user better understand if, for example, they have any hope of consistent behavior with their particular dataset and model space
- similarly, defining how exactly someone would figure out they need to increase CV folds would be very helpful
- citations in text and reference content at the end of the document are poorly formatted
- there are some sections that look like they maybe haven't been finished. an example would be the text block including `formally established (citation)`
- when I ran the section `Pooled TMLE Mixture Results`, I ended up with a CI range of [2.593, 2.735], which doesn't cover 3. obviously that's possible even if things are well-behaved, but please check to be sure things are all working as expected with your example
- folds: the `n_folds` parameter 5 and the folds in, for example, `plot_mixture_results`, are clearly referring to different things, as they do not match each other; but it's not clear from anything I saw in the package what the latter fold is referring to. please clarify
