### 2022-03-12 blind-contours/CVtreeMLE#5 @ghost
I'll add some assorted comments here, in addition to the above:

- first sentence contains broad generalization (e.g. `in most cases`) that I think weakens the presentation overall. that can be addressed later, and should involve specific examples
- direct reference to existing tools e.g. marginal models is essential here; certainly GLMs are a major offender here, but I think being as precise as possible strengthens the assertion that the package is needed
- I mention this in the vignette issue as well, but I think defining what is meant by `asymptotically` here is important: how can we evaluate if our data and models are appropriate for this package, or that we can expect reasonable convergence?
- the background section contains non-background content that is best included elsewhere or pared down
- how are guarantees in line 75 being proven? it seems like this in part relies on behavior of sl3. in any case, these guarantees should be elaborated upon
