### 2022-03-12 blind-contours/CVtreeMLE#5 @ghost
In terms of the review checklist, I think the two points `description [...] for a diverse, non-specialist audience` and `clearly states [...] who the target audience is` could both use further elaboration and expansion. The author is very clearly knowledgeable about the methodology in use here. I think the target audience would likely be users of standard marginal models (e.g. GLM), but observe that many such users are specifically using GLMs because they understand basic regression models but are perhaps not as comfortable with more advanced ML techniques. I think including some content that specifically reaches out to such users, if indeed they are part of the intended user set, would be most welcome.
