### 2022-03-12 blind-contours/CVtreeMLE#14 @cpalmer718
Linting
This isn't absolutely required, but it would be really nice to have this package passed through full linting. I know it can be a bit of a pain, but it can really help encourage best practices. As an example, I use [precommit](https://pre-commit.com/) with a config as follows:

```yaml
repos:
  ## R support
- repo: https://github.com/lorenzwalthert/precommit
  rev: v0.2.2
  hooks:
  - id: style-files
  - id: parsable-R
  - id: no-browser-statement
  - id: lintr
    verbose: true
  - id: roxygenize
  - id: deps-in-desc
  - id: use-tidy-description
```
