### 2022-03-12 #4181 @cpalmer718
Extended comments/issue links:

- practical test coverage around 4.47%: https://github.com/blind-contours/CVtreeMLE/issues/10
- contribution link broken: https://github.com/blind-contours/CVtreeMLE/issues/8
- I concur about function documentation: https://github.com/blind-contours/CVtreeMLE/issues/4
- package is not on CRAN, as mentioned above; but also the devel version instructions don't work either because there's no `devel` branch. `main` installation with install_github works fine though: https://github.com/blind-contours/CVtreeMLE/issues/11
- vignette needs reworking and editing: https://github.com/blind-contours/CVtreeMLE/issues/12
- additional comments about JOSS paper: https://github.com/blind-contours/CVtreeMLE/issues/5
- runtime performance guidelines: https://github.com/blind-contours/CVtreeMLE/issues/13
- linting: https://github.com/blind-contours/CVtreeMLE/issues/14
