### 2022-03-13 #4181 @osorensen
Thanks a lot for your review, @cpalmer718. Please notify us when these issues have been resolved, @blind-contours, and don't hesitate to reach out if I can be of any help.
