### 2022-06-10 a58875c
Update LICENSE.md
Updating license date.
