+++ README.md
@@ -846,13 +846,13 @@ where I was a biomedical big data fellow.
 
 ## License
 
-© 2017-2021 [David B. McCoy](https://davidmccoy.org)
+© 2017-2022 [David B. McCoy](https://davidmccoy.org)
 
 The contents of this repository are distributed under the MIT license.
 See below for details:
 
     MIT License
-    Copyright (c) 2017-2021 David B. McCoy
+    Copyright (c) 2017-2022 David B. McCoy
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"), to deal
     in the Software without restriction, including without limitation the rights
