### 2022-06-10 blind-contours/CVtreeMLE#2 @blind-contours
Hi @GaryBAYLOR, thank you -- the second license has been removed and the exiting MIT license has been updated to 2022 and matches the license listed in the READme.
