### 2022-06-10 blind-contours/CVtreeMLE#3 @blind-contours
Hi @GaryBAYLOR, thank you -- yes the CVtreeMLE package is currently not available on CRAN. This is because it depends currently on the sl3 ensemble machine learning package which is not on CRAN. The readme documentation has been updated. Future work will be done to get CVtreeMLE on CRAN, one approach for this is to instead build nonparametric estimators using the HAL9001 package (on CRAN) which is a nonparametric estimator with asymptoticly efficient properties.  In the meantime, CVtreeMLE must be downloaded from this repository.
