### 2022-06-13 blind-contours/CVtreeMLE#3 @GaryBAYLOR
If you intent to submit the package to CRAN after publishing the paper, can you also update the Availability Section in the paper to remove the statement that the package has been publicly available on CRAN, since it is not so far and it may take a certain amount of time to do that after publishing the paper. Instead, you can update the README file in the Github repo later to let the audience know it is on CRAN.
