### 2022-06-19 #4181 @osorensen
@openjournals/joss-eics I have not been able to get in touch with the submitting author @blind-contours, despite repeated attempts. The last response is on 17 March and none of the issues opened by the reviewers in the source repository appear to have been addressed. Could you please consider if this submission should be rejected, with an opportunity to resubmit? I have also searched extensively for the author's e-mail address, but not been able to find it.

Furthermore, the manuscript claims that "The CVtreeMLE package has been made publicly available both via GitHub and the Comprehensive R Archive Network." However, I am not able to find the package on the [Comprehensive R Archive Network](https://cran.r-project.org/).
