### 2022-06-19 #4181 @blind-contours
@osorensen -- sorry for the delay. I've started addressing the issues made by @cpalmer718. My email address is david_mccoy@berkeley.edu, I will add it to the readme and update the link to my bio page. I wasn't able to get the package on CRAN because the SL3 dependency is not on CRAN so I am removing that section of the manuscript and updating the READme. Thank you for your review of this package, I will be responding to issues this week.
