### 2022-06-19 c74d435
Edits to background based on issue 6
