### 2022-06-19 blind-contours/CVtreeMLE#6 @blind-contours
I have added additional sentences to clarify V-fold cross validation (in the targeted learning literature we call this process V-fold CV for data-adaptive target parameters).

I've changed the background section to be more intuitive with less jargon. The general issue is that in most cases we have one treatment/exposure so we know what the exposure is. In the case for mixtures, we need to define exposure given every combination of exposures in the mixture is not available. I've just detailed this in words rather than saying "high-dimensional" or "sparsity". Thank you for helping clarify this.
