### 2022-06-20 6dc818c
Updates to paper based on JOSS issue #5
