### 2022-06-20 blind-contours/CVtreeMLE#5 @blind-contours
@GaryBAYLOR I have made edits to the paper based on your comments. @cpalmer718 I have done some rewriting to reference existing methods and lay a quick but better foundation as to the methodological gap this method/software is filling. I've added some additional details as to asymptotic convergence and under what assumptions we have an efficient rate of convergence. I have also rewritten sections of the background.
