### 2022-06-21 bbe5d70
Update draft-pdf.yml
Can't figure out why the path doesn't work.
