+++ paper/paper.bib
@@ -1,3 +1,35 @@
+@article{Bobb2014,
+abstract = {Because humans are invariably exposed to complex chemical mixtures, estimating the health effects of multi-pollutant exposures is of critical concern in environmental epidemiology, and to regulatory agencies such as the U.S. Environmental Protection Agency. However, most health effects studies focus on single agents or consider simple two-way interaction models, in part because we lack the statistical methodology to more realistically capture the complexity of mixed exposures. We introduce Bayesian kernel machine regression (BKMR) as a new approach to study mixtures, in which the health outcome is regressed on a flexible function of the mixture (e.g. air pollution or toxic waste) components that is specified using a kernel function. In high-dimensional settings, a novel hierarchical variable selection approach is incorporated to identify important mixture components and account for the correlated structure of the mixture. Simulation studies demonstrate the success of BKMR in estimating the exposure-response function and in identifying the individual components of the mixture responsible for health effects. We demonstrate the features of the method through epidemiology and toxicology applications.},
+author = {Bobb, Jennifer F. and Valeri, Linda and {Claus Henn}, Birgit and Christiani, David C. and Wright, Robert O. and Mazumdar, Maitreyi and Godleski, John J. and Coull, Brent A.},
+doi = {10.1093/biostatistics/kxu058},
+file = {:Users/davidmccoy/Library/Application Support/Mendeley Desktop/Downloaded/Bobb et al. - 2014 - Bayesian kernel machine regression for estimating the health effects of multi-pollutant mixtures.pdf:pdf},
+issn = {14684357},
+journal = {Biostatistics},
+keywords = {Air pollution,Bayesian variable selection,Environmental health,Gaussian process regression,Metal mixtures.},
+number = {3},
+pages = {493--508},
+pmid = {25532525},
+title = {{Bayesian kernel machine regression for estimating the health effects of multi-pollutant mixtures}},
+volume = {16},
+year = {2014}
+}
+@article{Keil2019,
+abstract = {Background: Exposure mixtures frequently occur in data across many domains, particularly in the fields of environmental and nutritional epidemiology. Various strategies have arisen to answer questions about exposure mixtures, including methods such as weighted quantile sum (WQS) regression that estimate a joint effect of the mixture components. Objectives: We demonstrate a new approach to estimating the joint effects of a mixture: quantile g-computation. This approach combines the inferential simplicity of WQS regression with the flexibility of g-computation, a method of causal effect estimation. We use simulations to examine whether quantile g-computation and WQS regression can accurately and precisely estimate effects of mixtures in a variety of common scenarios. Methods: We examine the bias, confidence interval coverage, and bias-variance tradeoff of quantile g-computation and WQS regression, and how these quantities are impacted by the presence of non-causal exposures, exposure correlation, unmeasured confounding, and non-linearity of exposure effects. Results: Quantile g-computation, unlike WQS regression allows inference on mixture effects that is unbiased with appropriate confidence interval coverage at sample sizes typically encountered in epidemiologic studies, and when the assumptions of WQS regression are not met. Further, WQS regression can magnify bias from unmeasured confounding that might occur if important components of the mixture are omitted from analysis. Discussion: Unlike inferential approaches that examine effects of individual exposures, while holding other exposures constant, methods like quantile g-computation that can estimate the effect of a mixture are essential for understanding effects of potential public health actions that act on exposure sources. Our approach may serve to help bridge gaps between epidemiologic analysis and interventions such as regulations on industrial emissions or mining processes, dietary changes, or consumer behavioral changes that act on multiple exposures simultaneously.},
+archivePrefix = {arXiv},
+arxivId = {1902.04200},
[…]
