### 2022-06-21 892d465
Further paper edits based on issue #5
