### 2022-07-13 #4181 @blind-contours
Hi @osorensen, I have responded to issue #5 and #6 in regards to the JOSS paper submission. Generally, I have added details to elaborate on more abstruse terms and moved parts of different sections around to better match organization. Hopefully these edits meet the reviewers comments but please let me know if there are any other suggestions. The current paper.md is the most updated draft.
