### 2022-07-13 #4181 @blind-contours
Issues #1 and #3 have been resolved - this is regarding CVtreeMLE not being available on CRAN. I've simply removed this from the documentation and download is only available off github as the package currently relies on SL3 for ensemble machine learning. In the future, I can update code to use tidymodels or HAL9001 nonparametric estimators which will allow for CRAN submission but for now this is not available given SL3 is not on CRAN.
