### 2022-07-13 #4181 @blind-contours
I am currently cleaning the existing code to pass linting tests and updating documentation to meet reviewer comments #7, #4, and #14. I expect all issues to be resolved by August 1st.
