+++ R/utils.R
@@ -1,15 +1,20 @@
-#' Simply function that looks at variables used in the pre model for matches with the mixture variables and pulls them out to check for consistent sets of rules found over the iterative backfitting procedure
+#' @title  Pull rules out of results table of pre results
+#' @details Simply function that looks at variables used in the pre model
+#' for matches
+#'  with the mixture variables and pulls them out to check for consistent
+#'  sets of rules found over the iterative backfitting procedure
 #' @param x Row of a dataframe that contains output from the pre fit
-#' @param A Vector of characters indicating column names for the mixture variables
+#' @param a Vector of characters indicating column names for the mixture
+#' variables
 #' @return Vector of matches of mixture variables
 #' @export
 
-pull_out_rule_vars <- function(x, A) {
+pull_out_rule_vars <- function(x, a) {
   x <- x[3]
   if (x != "1") {
     x_split <- strsplit(x, " ")[[1]]
-    hits <- x_split[grep(paste(A, collapse = "|"), x_split)]
-    hits <- sort(hits)
+    hits <- x_split[grep(paste(a, collapse = "|"), x_split)]
+    hits <- sort(unique(hits))
     if (length(hits) == 1) {
       hits <- 0
     } else {
@@ -25,7 +30,6 @@ pull_out_rule_vars <- function(x, A) {
 ###################################################################
 #' @title Round rules found for easier reading
 #' @param rules Vector or rules
-
 #' @importFrom rlang .data
 round_rules <- function(rules) {
   rounded_rules <- list()
@@ -34,7 +38,8 @@ round_rules <- function(rules) {
     rule_split <- split_rules[[i]]
     for (j in seq(rule_split)) {
       element <- rule_split[[j]]
-      if (grepl("-*\\d+\\.*\\d*", element) == TRUE & grepl("[[:alpha:]]", element) == FALSE) {
+      if (grepl("-*\\d+\\.*\\d*", element) == TRUE &&
+          grepl("[[:alpha:]]", element) == FALSE) {
         element_round <- round(as.numeric(element), 3)
         rule_split[[j]] <- element_round
       }
@@ -49,10 +54,12 @@ round_rules <- function(rules) {
 
 ###################################################################
 #' @title Filter data based on fold
+#' @details  Filter data to only the training fold of interest during CV
 #' @param data Input data
 #' @param fold_k Current fold
-
 #' @importFrom rlang .data
+#'
+#' @export
 filter_rules <- function(data, fold_k) {
   data <- data %>%
     dplyr::filter(.data$fold == fold_k)
@@ -62,8 +69,9 @@ filter_rules <- function(data, fold_k) {
 ###################################################################
 #' @title Group by fold
 #' @param data Input data
-
 #' @importFrom rlang .data
+#'
+#' @export
 groupby_fold <- function(data) {
   data <- data %>%
     dplyr::group_by(.data$fold)
@@ -71,11 +79,13 @@ groupby_fold <- function(data) {
 }
 
 ###################################################################
-#' @title Filter mixture rules across the folds for only those that have the same variables and directions across all folds
+#' @title Filter mixture rules across the folds for only those that have the
+#' same variables and directions across all folds
 #' @param data Input data
 #' @param n_folds Number of folds in the CV
-
 #' @importFrom rlang .data
+#'
+#' @export
 filter_mixture_rules <- function(data, n_folds) {
   data <- data %>%
[…]
