### 2022-07-25 d2e90f6
Update R-CMD-check.yml
added xgboost to yml install.
