+++ README.Rmd
@@ -15,9 +15,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# `CVtreeMLE` <img src="man/figures/CVtreeMLE_sticker.png" 
-# height="300" align="right"/>
-
+# `CVtreeMLE` <img src="man/figures/CVtreeMLE_sticker.png height="300" align="right"/>
 <!-- badges: start -->
 [![R-CMD-check](https://github.com/blind-contours/CVtreeMLE/workflows/R-CMD-check/badge.svg)](https://github.com/blind-contours/CVtreeMLE/actions)
 [![Coverage Status](https://codecov.io/gh/blind-contours/CVtreeMLE/branch/main/graph/badge.svg?token=HJP5PYQSG4)](https://codecov.io/github/blind-contours/CVtreeMLE?branch=master)
@@ -73,14 +71,14 @@ for public policy and are asymptotically efficient.
 [@coyle2021sl3] to allow for ensemble machine learning to be leveraged 
 in the estimation procedure. `sl3` is used to 
 create ensemble machine learning estimators for the $Q$ and $g$ mechanisms for 
-the average treatment effect (ATE) target parameter and is also used in the 
+the average treatment effect (ATE) target parameter. `sl3`  is also used in the 
 iterative backfitting procedure. In the iterative
-backfitting procedure, for an enselbe of decision trees are fit on the full 
+backfitting procedure,  an ensemble of decision trees are fit on the full 
 mixture modeled 
 together, the [`pre` package](https://github.com/marjoleinF/pre)[@Fokkema2020a]
 is used to fit rule
-ensembles. In backfitting procedure to find quantiles in each mixture component
-individually, a Super Learner of decision trees generated from the [`partykit` package](http://partykit.r-forge.r-project.org/partykit/)[partykit2015] is 
+ensembles. In backfitting procedure to find thresholds in each mixture component
+individually, a Super Learner of decision trees generated from the [`partykit` package](http://partykit.r-forge.r-project.org/partykit/)[partykit2015] are 
 created. In each case, the goal is to find the best fitting decision tree 
 from which we extract decision tree rules, we then calculate the ATE for 
 these rules.
@@ -143,7 +141,7 @@ is highest in this region and where the size of the region is based on
 covariates. 
 
 
-To do this, use the `simulate_mixture_cube` function to generate simulated 
+To do this, use the `simulate_mixture_cube` function to generate simulated    
 data that represents ground-truth. Here, we create three continuous mixture 
 variables, $A$, that are correlated and baseline covariates, $W$, that are 
 potential confounders. Our outcome will be generated such that individuals 
@@ -190,24 +188,31 @@ We will now pass the simulated data, learners, and variable names for each
 node in $O = W,A,Y$ to the `CVtreeMLE` function:
 
 ```{r run simulation}
-set.seed(2333)
-
 ptm <- proc.time()
 
 sim_results <- CVtreeMLE(data = sim_data,
                          w = c("w", "w2"),
                          a = c(paste("M", seq(3), sep = "")),
                          y = "y",
-                         n_folds = 3,
+                         n_folds = 5,
+                         parallel_cv = TRUE,
+                         seed = 2333,
+                         parallel_type = "multi_session",
                          family = "gaussian",
-                         num_cores = 5)
+                         num_cores = 2)
 
 proc.time() - ptm
 ```
 
 Note that above, there are default estimators for all parameters if they are
 not passed to the function. Here we just use the out of the box estimators 
-that are defined in `utils_create_sls.R`
+that are defined in `utils_create_sls.R`. These estimators are chosen to be
+both non-parametric but also not too computationally demanding. Examples 
+of estimators used by default are random forest, xgboost, elastic net, and glms. 
+Users can also pass in their own custom stacks of learners. We also see here 
[…]
