+++ README.md
@@ -1,9 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# `CVtreeMLE` &lt;img src=“man/figures/CVtreeMLE\_sticker.png”
-
-# height=“300” align=“right”/&gt;
+# `CVtreeMLE` \<img src=“man/figures/CVtreeMLE_sticker.png height=”300” align=“right”/\>
 
 <!-- badges: start -->
 
@@ -49,10 +47,13 @@ data-adaptive parameters see (Hubbard, Kherad-Pajouh, and Van Der Laan
 `CVtreeMLE` uses data-adaptive parameters by implementing V-fold
 cross-validation (CV), that is, in 10-fold CV, the data is split 10
 times (folds), where 90% of the data is used to determine rules in a
-mixture, and the *g* (probability of being exposued estimator, P(A\|W))
-and *Q* (outcome estimator E(Y\|A,W)) estimators needed for the ATE.
-These rules and estimators created in training data are applied to the
-validation data in order to calculate the final ATE target parameter.
+mixture, and the
+![g](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;g "g")
+(probability of being exposued estimator, P(A\|W)) and
+![Q](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;Q "Q")
+(outcome estimator E(Y\|A,W)) estimators needed for the ATE. These rules
+and estimators created in training data are applied to the validation
+data in order to calculate the final ATE target parameter.
 
 In order to optimize the optimum bias-variance trade-off for our causal
 parameter of interest we use cross-validated targeted minimum loss based
@@ -69,16 +70,20 @@ asymptotically efficient.
 [`sl3`package](https://github.com/tlverse/sl3) (Coyle et al. 2021) to
 allow for ensemble machine learning to be leveraged in the estimation
 procedure. `sl3` is used to create ensemble machine learning estimators
-for the *Q* and *g* mechanisms for the average treatment effect (ATE)
-target parameter and is also used in the iterative backfitting
-procedure. In the iterative backfitting procedure, for an enselbe of
-decision trees are fit on the full mixture modeled together, the [`pre`
+for the
+![Q](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;Q "Q")
+and
+![g](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;g "g")
+mechanisms for the average treatment effect (ATE) target parameter.
+`sl3` is also used in the iterative backfitting procedure. In the
+iterative backfitting procedure, an ensemble of decision trees are fit
+on the full mixture modeled together, the [`pre`
 package](https://github.com/marjoleinF/pre)(Fokkema 2020) is used to fit
-rule ensembles. In backfitting procedure to find quantiles in each
+rule ensembles. In backfitting procedure to find thresholds in each
 mixture component individually, a Super Learner of decision trees
 generated from the [`partykit`
 package](http://partykit.r-forge.r-project.org/partykit/)\[partykit2015\]
-is created. In each case, the goal is to find the best fitting decision
+are created. In each case, the goal is to find the best fitting decision
 tree from which we extract decision tree rules, we then calculate the
[…]
