+++ doc/intro_CVtreeMLE.R
@@ -13,6 +13,7 @@ library(kableExtra)
 library(dplyr)
 
 seed <- 5454432
+set.seed(seed)
 
 ## ----simulate data, warning=FALSE---------------------------------------------
 NIEHS_data <- NIEHS_data_1
@@ -50,20 +51,21 @@ pooled_mixture_results <- NIEHS_results$`Pooled TMLE Mixture Results`
 
 pooled_mixture_results %>%
   dplyr::filter(Proportion_Folds == 1.0)  %>%
+  dplyr::arrange(desc(`Mixture ATE`)) %>%
   kableExtra::kbl(caption = "Pooled TMLE Mixture Results") %>%
   kableExtra::kable_classic(full_width = F, html_font = "Cambria")
 
 ## ----fold specific mixture results--------------------------------------------
 vfold_mixture_results <- NIEHS_results$`V-Specific Mix Results`
 
-vfold_mixture_results$X2X5 %>%
-  kableExtra::kbl(caption = "X2 and X5 Interaction") %>%
+vfold_mixture_results$X5X7 %>%
+  kableExtra::kbl(caption = "X5 and X7 Interaction") %>%
   kableExtra::kable_classic(full_width = F, html_font = "Cambria")
 
-
 ## ----plot sim_mixture_results, fig.height = 3, fig.width = 8------------------
-mixture_plots <- plot_mixture_results(v_intxn_results = NIEHS_results$`V-Specific Mix Results`,hjust = 0.8)
-mixture_plots$X1X5
+mixture_plots <- plot_mixture_results(
+  v_intxn_results = NIEHS_results$`V-Specific Mix Results`,hjust = 0.8)
+mixture_plots$X5X7
 
 ## ----marginal results---------------------------------------------------------
 pooled_marginal_results <- NIEHS_results$`Pooled TMLE Marginal Results`
@@ -78,5 +80,25 @@ pooled_marginal_refs<- NIEHS_results$`Pooled Marginal Refs`
 
 pooled_marginal_refs %>%
   kableExtra::kbl(caption = "Pooled Marginal Reference Regions") %>%
+  kableExtra::kable_classic(full_width = F, html_font = "Cambria")
+
+## ----marginal_results_example-------------------------------------------------
+pooled_marginal_results %>%
+  dplyr::filter(`Proportion in Fold` == 1)  %>%
+  dplyr::filter(var == "X1")  %>%
+  kableExtra::kbl(caption = "Pooled Marginal Results") %>%
+  kableExtra::kable_classic(full_width = F, html_font = "Cambria")
+
+## ----marginal refs example----------------------------------------------------
+pooled_marginal_refs %>%
+  dplyr::filter(`Variable Quantile` == "X1_1") %>%
+  kableExtra::kbl(caption = "Pooled Marginal Reference Regions X1") %>%
+  kableExtra::kable_classic(full_width = F, html_font = "Cambria")
+
+## ----fold marginal results----------------------------------------------------
+fold_marginal_results<- NIEHS_results$`V-Specific Marg Results`
+
+fold_marginal_results$`X1_2-X1_1` %>%
+  kableExtra::kbl(caption = "V-fold specific results for X1 regions 1 and 2") %>%
   kableExtra::kable_classic(full_width = F, html_font = "Cambria")
 
