+++ R/CVtreeMLE.R
@@ -163,7 +163,7 @@ CVtreeMLE <- function(w,
                       aw_stack = NULL,
                       a_stack = NULL,
                       n_folds,
-                      seed = set.seed(),
+                      seed = set.seed(6442),
                       family,
                       h_aw_trunc_lvl = 10,
                       parallel = TRUE,
