### 2022-07-27 blind-contours/CVtreeMLE#9 @blind-contours
This has been fixed. Issue was that seeds created in the parallelized section would override the seed created. I've switched everything to furrr::future_map now and made sure the seed is set in each function. Now - users pass a seed number and this is used in each function. The above test now gives the same results for each seed.
