### 2022-07-27 blind-contours/CVtreeMLE#13 @blind-contours
I've added runtime performance guidelines in the vignette. I've also added default learners for each parameter that are nonparametric but also run pretty quickly. Install and running all the tests will still take a while but for a standard dataset this should run on a modern local machine within a couple hours depending the dimensions and number of folds chosen. Proc time is given in each example so users can see how long it runs on some simulated data and real world type synthetic data.
