### 2022-07-27 blind-contours/CVtreeMLE#3 @blind-contours
I have edited the availability section to reflect that the package is currently not on CRAN.
