### 2022-07-27 blind-contours/CVtreeMLE#10 @blind-contours
I have included examples in the documentation that run an example of each major function. I've added and am still adding more testing. With the breast cancer test, our goal is to ensure that the output interactions are in the correct format - I've updated this test script accordingly. Other tests have been included to ensure we are getting, for a respective variable set, the rules that have the maximum explanatory on the outcome in the iterative backfitting procedure. I'm adding some additional tests for the estimators as well.
