### 2022-07-27 blind-contours/CVtreeMLE#14 @blind-contours
A majority of the package has been linted now to the style guide of the tlverse. There is still some things to do but for the most part the coding is aligned with best practices.
