### 2022-07-28 blind-contours/CVtreeMLE#8 @blind-contours
I have updated the brief explanation of the simulation and the respective "mixture cube". I've increased the nfolds to 5, originally it was smaller for easier testing of the package. I've added a real world example in the vignette and there is a breast cancer example in the tests. I've included discussion of ML via SL3 in the readme and vignette.
