### 2022-07-28 blind-contours/CVtreeMLE#7 @blind-contours
I've done a massive editing and via linting identified and removed all these areas.
