### 2022-07-28 blind-contours/CVtreeMLE#4 @blind-contours
Current `R CMD CHECK` passes with no notes, warnings or errors. All documentation has been updated. There are some areas I would like to give better detail on but all functions exported to the user have detailed documentation.
