### 2022-07-28 6adccba
New tests, update readme
