+++ README.Rmd
@@ -26,6 +26,7 @@ knitr::opts_chunk$set(
 [![MIT license](https://img.shields.io/badge/license-MIT-brightgreen.svg)](https://opensource.org/licenses/MIT)
 <!-- [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4070042.svg)](https://doi.org/10.5281/zenodo.4070042) -->
 <!-- [![DOI](https://joss.theoj.org/papers/10.21105/joss.02447/status.svg)](https://doi.org/10.21105/joss.02447) -->
+[![Codecov test coverage](https://codecov.io/gh/blind-contours/CVtreeMLE/branch/main/graph/badge.svg)](https://app.codecov.io/gh/blind-contours/CVtreeMLE?branch=main)
 <!-- badges: end -->
 
 > Efficient Estimation of the Causal Effects of Joint Exposure using Data 
@@ -151,7 +152,7 @@ individuals who are not exposed to this combination of exposure levels.
 ## Simulate Data
 
 ```{r simulate data, warning=FALSE}
-sim_data <- simulate_mixture_cube(n_obs = 500,
+sim_data <- simulate_mixture_cube(n_obs = 800,
                                   splits = c(0.99, 2.0, 2.5),
                                   mins = c(0, 0, 0),
                                   maxs = c(3, 4, 5),
