### 2022-07-28 #4181 @blind-contours
@osorensen et. al., I have made significant changes to the package in response to the JOSS issues page. I have given responses to each and made updates to the package accordingly. I'm still making some small changes but for the most part I believe the new updates are a major improvement in the package given the issues raised by reviewers. This is the first major package I have created so I really appreciate the thorough review. Please let me know what additional changes you would like to see after reviewing what I have already done. Thanks again - really appreciate the time and effort.
