### 2022-07-28 blind-contours/CVtreeMLE#12 @blind-contours
I need to add some description of defining how exactly someone would figure out they need to increase CV folds which can be included still given changes to the vignette.
