+++ sandbox/03_run_simulation.R
@@ -19,17 +19,17 @@ library(dplyr)
 devtools::load_all(here())
 
 # simulation parameters
-set.seed(7259)
+seed <- set.seed(7259)
 n_sim <- 2 # number of simulations
 n_obs <- (cumsum(rep(sqrt(40), 6))^2)[-1] # sample sizes at root-n scale
 # n_obs <- n_obs[1:3]
-truth <- c(1, 1, 1, 1, 1, 1, 6, 7)
-true_rule <- "M2 > 2.5 & M3 > 3.6"
+truth <- c(1, 1, 1, 1, 1, 1, 1, 7)
+true_rule <- "M1 > 1.0 & M2 > 2.5 & M3 > 3.5"
 
-source(here("sandbox", "01_setup_data.R"))
 source(here("sandbox", "02_fit_estimators.R"))
 
-P_0_data <- generate_P_0(truth = truth, n = 100000)
+P_0_data <- simulate_mixture_cube(n_obs = 100000,
+                                  subspace_assoc_strength_betas = truth)
 
 # perform simulation across sample sizes
 sim_results <- lapply(n_obs, function(sample_size) {
