+++ inst/REFERENCES.bib
@@ -1,37 +1,89 @@
-  @Article{partykit2015,
-    title = {{partykit}: A Modular Toolkit for Recursive Partytioning
-      in {R}},
-    author = {Torsten Hothorn and Achim Zeileis},
-    journal = {Journal of Machine Learning Research},
-    year = {2015},
-    volume = {16},
-    pages = {3905-3909},
-    url = {https://jmlr.org/papers/v16/hothorn15a.html},
-  }
-  @Article{Hothorn2006,
-    title = {Unbiased Recursive Partitioning: A Conditional Inference
-      Framework},
-    author = {Torsten Hothorn and Kurt Hornik and Achim Zeileis},
-    journal = {Journal of Computational and Graphical Statistics},
-    year = {2006},
-    volume = {15},
-    number = {3},
-    doi = {10.1198/106186006X133933},
-    pages = {651--674},
-  }
-  @Article{Zeileis2008,
-    title = {Model-Based Recursive Partitioning},
-    author = {Achim Zeileis and Torsten Hothorn and Kurt Hornik},
-    journal = {Journal of Computational and Graphical Statistics},
-    year = {2008},
-    volume = {17},
-    number = {2},
-    doi = {10.1198/106186008X319331},
-    pages = {492--514},
-  }
+@article{Athey2016a,
+abstract = {In this paper we propose methods for estimating heterogeneity in
+causal effects in experimental and observational studies and for conducting
+hypothesis tests about the magnitude of differences in treatment effects across
+subsets of the population. We provide a data-driven approach to partition the
+data into subpopulations that differ in the magnitude of their treatment
+effects. The approach enables the construction of valid confidence intervals
+for treatment effects, even with many covariates relative to the sample size,
+and without "sparsity" assumptions. We propose an "honest" approach to
+estimation, whereby one sample is used to construct the partition and another
+to estimate treatment effects for each subpopulation. Our approach builds on
+regression tree methods, modified to optimize for goodness of fit in treatment
+effects and to account for honest estimation. Our model selection criterion
+anticipates that bias will be eliminated by honest estimation and also accounts
+for the effect of making additional splits on the variance of treatment effect
+estimates within each subpopulation. We address the challenge that the
+"ground truth" for a causal effect is not observed for any individual unit,
+so that standard approaches to cross-validation must be modified. Through a
+simulation study, we show that for our preferred method honest estimation
+results in nominal coverage for 90% confidence intervals, whereas coverage
+ranges between 74% and 84% for nonhonest approaches. Honest estimation
+requires estimating the model with a smaller sample size; the cost in terms
+of mean squared error of treatment effects for our preferred method ranges
+between 7-22%.},
+archivePrefix = {arXiv},
+arxivId = {1504.01132},
+author = {Athey, Susan and Imbens, Guido},
+doi = {10.1073/pnas.1510489113},
+eprint = {1504.01132},
+file = {:Users/davidmccoy/Downloads/pnas.1510489113 (1).pdf:pdf},
+issn = {10916490},
+journal = {Proceedings of the National Academy of Sciences of the United States of America},
+keywords = {Causal inference,Cross-validation,Heterogeneous treatment effects,Potential outcomes,Supervised machine learning},
+number = {27},
[…]
