### 2022-08-21 #4181 @osorensen
👋 @cpalmer718 and @GaryBAYLOR, could you please consider whether the updates made by @blind-contours address your points raised in the reviews, and update your review checklists accordingly?
