### 2022-08-21 blind-contours/CVtreeMLE#14 @ghost
Great. The remainder of the formatting warnings can be deferred until such time as you look to post on CRAN or something like that.
