### 2022-08-21 #4181 @cpalmer718
Hi @osorensen @blind-contours the package is much improved, and substantial effort has been put into addressing my comments. I'll summarize the following remaining notes:

- the only potentially tough one is that `@import {package}` statements are causing huge bloat to the package namespace. I'd recommend removing all of them, but if you can't avoid it for sl3 for some reason, at least remove it for partykit and especially ggplot2
- the contributions link in the readme still points to nothing
- the paper needs minor editing:
  - there are subject/verb agreement issues (e.g. `Newer mixture methods [...] is biased by`), specifically in the summary
  - I'd recommend splitting the summary text into at least two paragraphs. content is good but it's hard to read as a big block
- markdown in both readme and the vignette has some formatting issues:
  - vignette header text rendered as `## Marginal Results`
  - assorted links and references in readme render strangely
  - license text in readme and in quote block and in actual license file all differ by year
  - one instance of the license seems to preemptively claim 2023, which isn't a common convention
- consolidate .gitignore files into a single, top-level entry
- remove .DS_Store from repo

These are all quite minor, and beyond that I am content.
