### 2022-08-21 #4181 @osorensen
On the occasion, I'd also like to ask @blind-contours about the prospects of submitting this package to CRAN. Although not a strict requirement for publishing in JOSS, it definitely would be an advantage.
