### 2022-08-23 7985c2c
Updates based on new JOSS review, allow users to pass one mixture variable.
