### 2022-08-23 7a0237e
Remove 2D simulation because of bloat.
