### 2022-08-25 #4181 @blind-contours
Thank you @cpalmer718! As per your first point, I'm looking to reduce the bloat. I am still learning the intricacies of namespaces - are there particular packages you see that are only used via library()? I'll work on your additional comments. Really appreciate your thorough review! It's been a great learning experience.
