### 2022-08-25 #4181 @blind-contours
I would like to get this package on CRAN. The major issue is the ensemble machine learning used to estimate the data adaptive parameter and the nuisance parameters. Because SL3 - the ensemble machine learning package I use, is not on CRAN, and I can't really make it optional while keeping the statistical properties of the estimates, I would need to use a package that is on CRAN and does a type of Super Learning. I believe I can substitute SL3 for tidymodels. So my plan is, once this package is on JOSS (hopefully) to create a new branch for development of transferring the machine learning functions to use tidymodels - this would require a lot of restructuring of the package but would make it available on CRAN.
