### 2022-08-26 #4181 @osorensen
Another example I found while browsing source code. In `simulate_mixture_cube.R` you have the line

```
    dplyr::mutate(rcat = Hmisc::rMultinom(probs, 1))
 ```

but you also write

```
#' @importFrom Hmisc rMultinom
```

Again, since you use the qualifier `Hmisc::`, the `@importFrom` should be removed here.
