+++ R/simulate_mixture_cube.R
@@ -42,7 +42,6 @@
 #' @importFrom dplyr mutate
 #' @importFrom MASS mvrnorm
 #' @importFrom rlang :=
-#' @importFrom Hmisc rMultinom
 #' @export
 
 
