### 2022-08-27 blind-contours/CVtreeMLE#17 @GaryBAYLOR
Issue with NAMESPACE: extra space
Leave one space instead of multiple spaces between `#'` and `@export` when using package `roxygen2` to label a function to be exported. Otherwise the function will not appear in NAMESPACE file. I checked the file `bound.R` in which I see two spaces between `#'` and `@export` for all  functions in the file: `bound_precision()`, `bound_propensity()`, `scale_to_unit()`,  and `scale_to_original()`. For example,
```r
#' @return A \code{numeric} vector of the same length as \code{vals}, where
#'  the returned values are bounded to machine precision. This is intended to
#'  avoid numerical instability issues.
#'  @export
```
You can see those functions are not actually exported in NAMESPACE file. Please check if all function you intend to export are correctly generated in NAMESPACE file after running `roxygen2::roxygenise()`
