### 2022-08-27 blind-contours/CVtreeMLE#18 @GaryBAYLOR
Issue with package manual: incomplete section Value
A large number of function miss the description for the section **Value**. This section simply contains the following sentence
```r
#' @return Rules object. TODO: add more detail here.
```

The functions having the above issue include
```r
calc_marginal_ate
calc_mixtures_ate
calc_v_fold_marginal_ate
calc_v_fold_mixtures_ate
common_mixture_rules
est_comb_exposure
est_marg_nuisance_params
evaluate_marginal_rules
evaluate_mixture_rules
find_common_marginal_rules
fit_mix_rule_backfitting
meta_mix_results
simulate_mixture_cube
```
Please add complete description in this section so the users know the details of returned value of a function.
