### 2022-08-28 blind-contours/CVtreeMLE#19 @GaryBAYLOR
Multiple issues detected when running `devtools::check()`
Issues when running `devtools::check()`:

* Fix the duplicate chunk label issue in the vignette described in the following error message when running `devtools::check()`. This should be easy to fix.
```r
E  creating vignettes (10.2s)
   --- re-building ‘intro_CVtreeMLE.Rmd’ using rmarkdown
   Error: processing vignette 'intro_CVtreeMLE.Rmd' failed with diagnostics:
   Duplicate chunk label 'plot sim_mixture_results', which has been used for the chunk:
   mixture_plots <- plot_mixture_results(
     v_intxn_results = niehs_results$`V-Specific Mix Results`,
     hjust = 0.8)
   mixture_plots$X5X7
   Error: tangling vignette 'intro_CVtreeMLE.Rmd' failed with diagnostics:
   Duplicate chunk label 'plot sim_mixture_results', which has been used for the chunk:
   mixture_plots <- plot_mixture_results(
     v_intxn_results = niehs_results$`V-Specific Mix Results`,
     hjust = 0.8)
   mixture_plots$X5X7
   --- failed re-building ‘intro_CVtreeMLE.Rmd’
```

* Check and fix the following issue when running `devtools::check()` which is also related to vignette
```r
E  creating vignettes (34.4s)
   --- re-building ‘intro_CVtreeMLE.Rmd’ using rmarkdown
   Quitting from lines 114-127 (intro_CVtreeMLE.Rmd)
   Error: processing vignette 'intro_CVtreeMLE.Rmd' failed with diagnostics:
   could not find function "make_sl3_Task"
   --- failed re-building ‘intro_CVtreeMLE.Rmd’

   SUMMARY: processing the following file failed:
     ‘intro_CVtreeMLE.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error in get_parentpid() : attempt to apply non-function
Calls: <Anonymous> ... rstudio_stdout -> rstudio_detect -> detect_new -> get_parentpid
Execution halted
```
