### 2022-08-28 blind-contours/CVtreeMLE#4 @GaryBAYLOR
* A large number of functions exported to users miss the documentation for returned value section, see https://github.com/blind-contours/CVtreeMLE/issues/18
