### 2022-08-28 blind-contours/CVtreeMLE#5 @GaryBAYLOR
@blind-contours Looks good. Just one very minor comment, be consistent in writing `nonparametric`, either `non-parameter` or `nonparametric` is okay (line 18, 19, 30, 44)

@cpalmer718 I'm good on my comments. Please feel free to close the issue when the questions in your comments are addressed.
