### 2022-08-28 #4181 @GaryBAYLOR
@blind-contours I have updated some issues as well as adding a few new issues I detected. I will take time to look at issue #6 and #8 I raised previously given you have updated them.

* updated issues https://github.com/blind-contours/CVtreeMLE/issues/4,  https://github.com/blind-contours/CVtreeMLE/issues/5
* added a few new issues: https://github.com/blind-contours/CVtreeMLE/issues/17, https://github.com/blind-contours/CVtreeMLE/issues/18, https://github.com/blind-contours/CVtreeMLE/issues/19, https://github.com/blind-contours/CVtreeMLE/issues/20
