### 2022-08-29 blind-contours/CVtreeMLE#18 @blind-contours
@GaryBAYLOR - thanks for catching this. I've updated all the return values for these functions. Originally, these functions were not exported to the user. However, because now I am using furrr to parallelize using future_map, this requires passing the function in furrr_options so the functions for CVtreeMLE are available during parallelization. It seems that only exported functions are available so I had to export all the functions. For the advanced user this is likely more helpful as they can better trace what's going on under the hood.
