### 2022-08-29 blind-contours/CVtreeMLE#20 @blind-contours
@GaryBAYLOR - sorry about this. I believe this is because I removed sl3 and the other algorithm packages from the namespace and forgot to include these as libraries in the test scripts. I'll update and check.
