### 2022-08-30 blind-contours/CVtreeMLE#17 @blind-contours
@GaryBAYLOR  - thanks for spotting this. I have gone through all the functions and checked for double spaces before export. All export functions should now appear in the namespace.
