+++ R/format_fold_results.R
@@ -1,18 +1,0 @@
-
-format_fold_results <- function(marginal_rules,
-                                mix_rules,
-                                mix_combo_data,
-                                marg_combo_data) {
-
-
-
-
-
-
-
-
-  return(list(
-    "marginal_rules" = marginal_rules,
-    "mix_rules" = mix_rules
-  ))
-}
