### 2022-09-03 blind-contours/CVtreeMLE#20 @blind-contours
@GaryBAYLOR - Running check now should return only 3 warnings that I'm trying to resolve. But there should be notes or errors now.
