### 2022-09-03 blind-contours/CVtreeMLE#19 @blind-contours
Check should pass with only some small warnings that I'm trying to resolve. Vignette has been updated with new visualizations of the fold specific trees, more grounded description of goals, examples of reporting results etc. Duplicate chunk issues have been resolved.
