### 2022-09-03 blind-contours/CVtreeMLE#5 @blind-contours
I've updated the md document to be consistent in using non-parametric.
