### 2022-09-21 5f2db2e
sbatch script for 2d sim
