+++ sandbox/simulate_2d_data.R
@@ -5,7 +5,6 @@ library(readxl)
 library(tidyr)
 library(MASS)
 library(stringr)
-library(plot3D)
 library(dplyr)
 library(sl3)
 library(pre)
