### 2022-09-28 242d000
Splitting sim into quarters
