+++ sandbox/03_run_2d_simulation_r2.R
@@ -96,4 +96,4 @@ sim_results <- lapply(n_obs, function(sample_size) {
 names(sim_results) <- paste("n", n_obs, sep = "_")
 timestamp <- str_replace_all(Sys.time(), " ", "_")
 saveRDS(object = sim_results,
-        file = here("sandbox/data", paste0("CVtreeMLE_", "run_1", ".rds")))
+        file = here("sandbox/data", paste0("CVtreeMLE_", "run_2", ".rds")))
