+++ R/CVtreeMLE.R
@@ -423,7 +423,6 @@ CVtreeMLE <- function(w,
   }
 
   results <- furrr::future_map(unique(data$folds), function(fold_k) {
-
     # for (fold_k in unique(data$folds)) {
 
     at <- data[data$folds != fold_k, ]
@@ -539,6 +538,7 @@ CVtreeMLE <- function(w,
       "marg combo data")
 
     results_list
+
 
   }, .options = furrr::furrr_options(seed = seed, packages = c("CVtreeMLE",
                                                               "sl3")))
