+++ vignettes/intro_CVtreeMLE.Rmd
@@ -128,7 +128,7 @@ niehs_results <- CVtreeMLE(data = as.data.frame(niehs_data),
                          w = c("Z", "Z2", "Z3"),
                          a = c(paste("X", seq(7), sep = "")),
                          y = "Y",
-                         n_folds = 5,
+                         n_folds = 10,
                          seed = seed,
                          parallel_cv = TRUE,
                          family = "gaussian",
