### 2022-10-02 add8c73
Setting seed per iteration
