### 2022-10-02 483c2d6
Manual sample sizes.
