### 2022-10-03 7f0e5f6
NA in exposure
