### 2022-10-03 b89aa28
Just filtering for exposure combination that don't exist given covariance.
