### 2022-10-03 4236be1
Removing comparison to truth in simulation given it's a guess right now.
