### 2022-10-03 660b79b
Remove true rule
