### 2022-10-08 4314f7d
Updated family parameter, including now the direction parameter to grab the max or min rule from the ensemble.
