+++ man/fit_mix_rule_backfitting.Rd
@@ -74,6 +74,7 @@ example_output <- fit_mix_rule_backfitting(at = data,
                                             a = mix_comps,
                                             w = W,
                                             y = "y_scaled",
+                                            direction = "positive",
                                             w_stack = w_stack,
                                             fold = 1,
                                             max_iter = 1,
