### 2022-10-10 8cb1655
Added true ATE and rule to output.
