+++ tests/testthat/test_mixture_thresholds_eight_beta.R
@@ -3,6 +3,7 @@ library(testthat)
 library(sl3)
 library(pre)
 library(partykit)
+library(tidyr)
 
 
 seed <- 6442
@@ -15,7 +16,7 @@ sls <- create_sls()
 w_stack <- sls$W_stack
 tree_stack <- sls$A_stack
 mix_comps <- c("M1", "M2", "M3")
-w <- c("w", "w2")
+w <- c("age", "sex", "bmi")
 data$y_scaled <- data$y
 
 example_output <- fit_mix_rule_backfitting(at = data,
