+++ sandbox/03_run_2d_simulation_r1.R
@@ -15,7 +15,7 @@ Sys.unsetenv("GITHUB_PAT")
 devtools::install(upgrade = "never")
 
 # simulation parameters
-n_sim <- 10 # number of simulations
+n_sim <- 20 # number of simulations
 n_obs <- c(200, 350, 500, 750, 1000, 1500, 2000, 3000)
 
 # Establish globals ---------------------------
