### 2022-10-10 462b0c6
Update all sims to 25 with 4 fold CV
