+++ tests/testthat/test_mixture_thresholds_one_beta.R
@@ -19,6 +19,7 @@ example_output <- fit_mix_rule_backfitting(at = data,
                                              a = mix_comps,
                                              w = w,
                                              y = "y_scaled",
+                                             direction = "positive",
                                              w_stack = w_stack,
                                              fold = 1,
                                              max_iter = 1,
