+++ sandbox/03_run_2d_simulation_r2.R
@@ -15,26 +15,13 @@ Sys.unsetenv("GITHUB_PAT")
 devtools::install(upgrade = "never")
 
 # simulation parameters
-seed <- set.seed(7259)
-n_sim <- 5 # number of simulations
-n_obs <- (cumsum(rep(sqrt(40), 7))^2)[-1] # sample sizes at root-n scale
-# n_obs <- n_obs[1:3]
-true_rule <- "m1 > 4.15 & m2 > 5.17"
-
+n_sim <- 25 # number of simulations
+n_obs <- c(200, 350, 500, 750, 1000, 1500, 2000, 3000)
 
 # Establish globals ---------------------------
 
 # number of observations to create
 n <- 500000
-
-# means of the exposures
-mu <- c(1,2)
-
-# covariance of the exposures
-sigma <- matrix(c(1, 0.1,
-                  0.1, 1),
-                nrow = 2,
-                ncol = 2)
 
 # number of cuts per exposure
 n_cuts <- 5
@@ -49,59 +36,81 @@ c_matrix <- matrix(c(0.2,0.3,0.2,0.5),
                    ncol  = 2,
                    nrow = 2)
 
-# Generate exposures -----------------
-
-exposure_results <- gen_exposures(mu = mu,
-                                  sigma = sigma,
-                                  n = 5000000,
-                                  n_cuts)
-
-exposures <- exposure_results$exposures_labeled
-
 # Generate simulated data -----------------
 P_0_sim <- gen_covariates(n) %>% # gen covariates
   gen_multinom(exposure_grid = exposure_grid) %>% # assign obs to exposure cube
-  assign_exposures(exposure_grid = exposure_grid,
-                   exposures = exposures) %>% # put continuous exposure in cube
   assign_outcomes(exposure_grid = exposure_grid,
                   c_matrix = c_matrix)  # assign outcome based on cube
 
 P_0_data <- P_0_sim$data
+P_0_data_filt <- P_0_data[complete.cases(P_0_data) ,]
 
+# Calculate Empirical Truth -------------------------
+
+data <- P_0_data_filt
+m1_seq <- seq(1:5)
+m2_seq <- seq(1:5)
+m1_dir <- c(">=", "<=")
+m2_dir <- c(">=", "<=")
+
+result_list <- list()
+iter <- 1
+for (i in m1_seq) {
+  for (j in m2_seq) {
+    for (k in m1_dir) {
+      for (l in m2_dir) {
+        rule <- (paste("region1 ",k, i, "&", "region2", l, j))
+        ate_res <- calc_empir_truth(data, rule)
+        result_list[[iter]] <- c(rule, ate_res)
+        iter <- iter + 1
+
+      }
+    }
+  }
+}
+
+result_df <- as.data.frame(do.call(rbind,result_list))
+colnames(result_df) <- c("Rule", "EY_region", "EY_compl", "ATE")
+result_df$ATE <- as.numeric(result_df$ATE)
+p0_max_ate <- result_df[which.max(result_df$ATE),]
+true_rule <- p0_max_ate$Rule
+true_ate <- p0_max_ate$ATE
 # perform simulation across sample sizes
-sim_results <- lapply(n_obs, function(sample_size) {
+sim_results_df <- data.frame()
+
+for (sample_size in n_obs) {
   # get results in parallel
   results <- list()
 
   for(this_iter in seq_len(n_sim)) {
 
-    data_sim <-  P_0_data %>%
+    seed <- sample(1:10000,1)
+    set.seed(seed)
+
+    data_sim <-  P_0_data_filt %>%
       slice_sample(n = sample_size)
 
-    # print(mean(data_sim$outcome_obs))
-
-    # data_sim <- P_0_data %>%
[…]
