+++ tests/testthat/test_mixture_thresholds_eight_beta.R
@@ -23,6 +23,7 @@ example_output <- fit_mix_rule_backfitting(at = data,
                                              a = mix_comps,
                                              w = w,
                                              y = "y_scaled",
+                                             direction = "positive",
                                              w_stack = w_stack,
                                              fold = 1,
                                              max_iter = 2,
