### 2022-10-10 dd1dcce
Ensure that at least one observation has the ground-truth rule so that we can get confusion table metrics in sims.
