+++ sandbox/04_make_plots.R
@@ -50,32 +50,53 @@ sim_statistics_long <- sim_statistics %>%
   tidyr::gather(statistic, value, -c(n_obs))
 
 
-make_sim_statistics_plot <- function(sim_statistics_long, stats) {
+make_sim_statistics_plot <- function(sim_statistics_long, stats, labels,) {
   filtered_sim_stats <- sim_statistics_long %>%
     filter(
       statistic %in% stats
     )
 
   ggplot(filtered_sim_stats, aes(n_obs, value)) +
-    geom_point() + geom_line(linetype = "dashed") +
-    facet_wrap("statistic", scales = "free_y") +
+    geom_point(color = color) +
+    geom_line(color = color, size = 1) +
+    geom_point(color= color) +
+    facet_wrap("statistic",
+               scales = "free_y",
+               labeller = as_labeller(labels)) +
     theme_bw() +
     theme(
       axis.text.x = element_text(angle=45, vjust=1, hjust=1)
     ) +
-    scale_x_sqrt(breaks=n_obs)
+    scale_x_sqrt(breaks=n_obs)  + labs(x = "Number Observations", title = "Bias Measures")
+
 }
 
+plot_labels <- c(
+  "abs_true_bias" = "Absolute Bias Compared to Truth",
+  "abs_da_bias" = "Absolute Bias Compared to Data-Adaptive Truth",
+  "sqrt_n_abs_true_bias" = "Absolute Bias Compared to Truth Scaled by Root N",
+  "sqrt_n_abs_da_bias" = "Absolute Bias Compared to Data-Adaptive Truth Scaled by Root N"
+)
 
 CVtreeMLE_stats_plot <- make_sim_statistics_plot(
   sim_statistics_long,
   stats = c("abs_true_bias", "abs_da_bias",
-            "sqrt_n_abs_true_bias", "sqrt_n_abs_da_bias")
+            "sqrt_n_abs_true_bias", "sqrt_n_abs_da_bias"),
+  labels = plot_labels
+)
+
+plot_labels <- c(
+  "Mix_found" = "Proportion Mixture Rule Found",
+  "True_pos" = "Rule Indicator True Positive Compared to Ground-Truth",
+  "True_neg" = "Rule Indicator True Negative Compared to Ground-Truth",
+  "False_pos" = "Rule Indicator False Positive Compared to Ground-Truth",
+  "False_neg" = "Rule Indicator False Negative Compared to Ground-Truth"
 )
 
 CVtreeMLE_rule_stats_plot <- make_sim_statistics_plot(
   sim_statistics_long,
-  stats = c("Mix_found", "True_pos", "True_neg", "False_pos","False_neg")
+  stats = c("Mix_found", "True_pos", "True_neg", "False_pos","False_neg"),
+  color = "orange"
 )
 
 CVtreeMLE_rule_stats_plot <- make_sim_statistics_plot(
