### 2022-10-11 1626274
Updating simulations with different estimates based on pooled bias results
