+++ sandbox/03_run_2d_simulation_r1.R
@@ -15,7 +15,7 @@ source(here("sandbox", "simulate_2d_data.R"))
 # devtools::install(upgrade = "never")
 
 # simulation parameters
-n_sim <- 25 # number of simulations
+n_sim <- 5 # number of simulations
 n_obs <- c(200, 350, 500, 750, 1000, 1500, 2000, 3000)
 
 # Establish globals ---------------------------
