+++ sandbox/02_fit_estimators.R
@@ -12,7 +12,7 @@ fit_estimators <- function(data,
                            w = covars,
                            a = exposures,
                            y = outcome,
-                           n_folds = 4,
+                           n_folds = 10,
                            num_cores = 20,
                            family = "continuous",
                            direction = "positive",
@@ -23,84 +23,135 @@ fit_estimators <- function(data,
   max_ate_index <- which.max(abs(
     sim_results$`Pooled TMLE Mixture Results`$`Mixture ATE`))
 
-  if (length(max_ate_index) == 1) {
-    mixure_found_ind <- 1
-  } else{
-    mixure_found_ind <- 0
-  }
+  ## tmle pooled results ------------------------
+  tmle_pooled_mixture_results <- sim_results$`Pooled TMLE Mixture Results`[max_ate_index,]
+  tmle_pooled_lower <- tmle_pooled_mixture_results$`Lower CI`
+  tmle_pooled_upper <- tmle_pooled_mixture_results$`Upper CI`
+  tmle_pooled_ate <- tmle_pooled_mixture_results$`Mixture ATE`
+  tmle_pooled_mix_decisions <- tmle_pooled_mixture_results$Union_Rule
 
-  if (mixure_found_ind == 1) {
-    mixture_results <- sim_results$`Pooled TMLE Mixture Results`[max_ate_index,]
-    lower <- mixture_results$`Lower CI`
-    upper <- mixture_results$`Upper CI`
-    ate <- mixture_results$`Mixture ATE`
-    mix_decisions <- mixture_results$Union_Rule
+  ## tmle v-specific results ------------------------
+  tmle_v_mixture_results <- sim_results$`V-Specific Mix Results`$`region1-region2`
+  tmle_v_fold_mixture_results <- tmle_v_mixture_results[tmle_v_mixture_results$fold != "Pooled",]
+  v_spec_mean_ate <- mean(tmle_v_fold_mixture_results$ate)
+  v_spec_mean_lower <- mean(tmle_v_fold_mixture_results$lower_ci)
+  v_spec_mean_upper <- mean(tmle_v_fold_mixture_results$upper_ci)
 
-    DA_true_results <- calc_empir_truth(P_0_data, mix_decisions)
-    DA_P0_truth <- DA_true_results[3]
+  ## pooled results ------------------------
+  v_pooled_mixture_results <- tmle_v_mixture_results[tmle_v_mixture_results$fold == "Pooled",]
+  v_pooled_ate <- v_pooled_mixture_results$ate
+  v_pooled_lower <- v_pooled_mixture_results$lower_ci
+  v_pooled_upper <- v_pooled_mixture_results$upper_ci
 
-    rule_binary <- data %>%
-      dplyr::transmute(A = ifelse(eval(parse(text = mix_decisions)), 1, 0))
 
-    rule_applied_P_0 <- P_0_data %>%
-      dplyr::mutate(A = ifelse(eval(parse(text = mix_decisions)), 1, 0))
+  ate_results <- list("tmle_pooled_ate" = tmle_pooled_ate,
+                      "v_spec_mean_ate" = v_spec_mean_ate,
+                      "v_pooled_ate" = v_pooled_ate)
 
-    true_rule_binary <- data %>%
-      dplyr::mutate(A = ifelse(eval(parse(text = true_rule)), 1, 0))
+  lower_ci_results <-  list("tmle_pooled_lower" = tmle_pooled_lower,
+                            "v_spec_mean_lower" = v_spec_mean_lower,
+                            "v_pooled_lower" = v_pooled_lower)
 
-    DA_rule_bias <- ate - DA_P0_truth
-    bias <- true_ate - ate
+  upper_ci_results <-  list("tmle_pooled_upper" = tmle_pooled_upper,
[…]
