### 2022-10-14 ab38a9f
Parallelization in vignette was breaking the Linux check.
