+++ vignettes/intro_CVtreeMLE.Rmd
@@ -130,7 +130,7 @@ niehs_results <- CVtreeMLE(data = as.data.frame(niehs_data),
                          y = "Y",
                          n_folds = 10,
                          seed = seed,
-                         parallel_cv = TRUE,
+                         parallel_cv = FALSE,
                          family = "continuous",
                          num_cores = 7,
                          max_iter = 1)
