### 2022-10-14 ad947fe
Removed outcome scaling.
