+++ R/calc_v_fold_mixture_ate.R
@@ -58,28 +58,28 @@ calc_v_fold_mixtures_ate <- function(input_mix_rules,
         qbar_0w = mix_data$qbar_0w
       )
 
-      qbar_aw_star <- flux_results$qbar_aw_star
-      qbar_1w_star <- flux_results$qbar_1w_star
-      qbar_0w_star <- flux_results$qbar_0w_star
+      mix_data$qbar_aw_star <- flux_results$qbar_aw_star
+      mix_data$qbar_1w_star <- flux_results$qbar_1w_star
+      mix_data$qbar_0w_star <- flux_results$qbar_0w_star
 
-      ## back-scale Y
-      mix_data$qbar_aw_star <- scale_to_original(
-        scaled_vals = qbar_aw_star,
-        max_orig = max(mix_data[outcome]),
-        min(mix_data[outcome])
-      )
-
-      mix_data$qbar_0w_star <- scale_to_original(
-        scaled_vals = qbar_0w_star,
-        max_orig = max(mix_data[outcome]),
-        min(mix_data[outcome])
-      )
-
-      mix_data$qbar_1w_star <- scale_to_original(
-        scaled_vals = qbar_1w_star,
-        max_orig = max(mix_data[outcome]),
-        min(mix_data[outcome])
-      )
+      # ## back-scale Y
+      # mix_data$qbar_aw_star <- scale_to_original(
+      #   scaled_vals = qbar_aw_star,
+      #   max_orig = max(mix_data[outcome]),
+      #   min(mix_data[outcome])
+      # )
+      #
+      # mix_data$qbar_0w_star <- scale_to_original(
+      #   scaled_vals = qbar_0w_star,
+      #   max_orig = max(mix_data[outcome]),
+      #   min(mix_data[outcome])
+      # )
+      #
+      # mix_data$qbar_1w_star <- scale_to_original(
+      #   scaled_vals = qbar_1w_star,
+      #   max_orig = max(mix_data[outcome]),
+      #   min(mix_data[outcome])
+      # )
 
       ate_results <- calc_ate_estimates(
         data = mix_data,
