+++ sandbox/02_fit_estimators.R
@@ -12,7 +12,7 @@ fit_estimators <- function(data,
                            w = covars,
                            a = exposures,
                            y = outcome,
-                           n_folds = 10,
+                           n_folds = 2,
                            num_cores = 20,
                            family = "continuous",
                            direction = "positive",
