### 2022-10-14 a2a4658
Updating results around using unscaled y and changing funcitons accordingly.
