+++ R/fit_iterative_mix_rule_backfitting.R
@@ -30,7 +30,6 @@
 #'   }
 #' @examples
 #' data <- simulate_mixture_cube()
-#' data$y_scaled <- data$y
 #' mix_comps <- c("M1", "M2", "M3")
 #' W <- c("age", "sex", "bmi")
 #' sls <- create_sls()
@@ -39,7 +38,7 @@
 #' example_output <- fit_mix_rule_backfitting(at = data,
 #'                                             a = mix_comps,
 #'                                             w = W,
-#'                                             y = "y_scaled",
+#'                                             y = "y",
 #'                                             direction = "positive",
 #'                                             w_stack = w_stack,
 #'                                             fold = 1,
