### 2022-10-17 690a10c
Make marginals optional based on user preference.
