### 2022-10-18 afcf096
Empty DF when not fitting margs
