### 2022-10-18 418dd41
Updating run files to 20 iterations for server with marginal run updates.
