+++ sandbox/04_make_plots.R
@@ -37,12 +37,12 @@ sim_statistics <- sim_results %>%
     tmle_pooled_sd = sd(tmle_pooled_ate),
     v_spec_mean_sd = sd(v_spec_mean_ate),
     v_pooled_mean_sd = sd(v_pooled_ate),
-    tmle_pooled_da_mse = tmle_pooled_da_bias^2 + tmle_pooled_sd,
-    tmle_pooled_gt_mse = tmle_pooled_gt_bias^2 + tmle_pooled_sd,
-    v_spec_da_mean_mse = v_spec_da_mean_bias^2 + v_spec_mean_sd,
-    v_spec_gt_mean_mse = v_spec_gt_mean_bias^2 + v_spec_mean_sd,
-    v_pooled_da_mse = v_pooled_da_bias^2 + v_pooled_mean_sd,
-    v_pooled_gt_mse = v_pooled_gt_bias^2 + v_pooled_mean_sd,
+    tmle_pooled_da_mse = tmle_pooled_da_bias^2 + tmle_pooled_sd^2,
+    tmle_pooled_gt_mse = tmle_pooled_gt_bias^2 + tmle_pooled_sd^2,
+    v_spec_da_mean_mse = v_spec_da_mean_bias^2 + v_spec_mean_sd^2,
+    v_spec_gt_mean_mse = v_spec_gt_mean_bias^2 + v_spec_mean_sd^2,
+    v_pooled_da_mse = v_pooled_da_bias^2 + v_pooled_mean_sd^2,
+    v_pooled_gt_mse = v_pooled_gt_bias^2 + v_pooled_mean_sd^2,
     tmle_pooled_gt_coverage = mean(tmle_pooled_gt_coverage),
     tmle_pooled_da_coverage = mean(tmle_pooled_da_coverage),
     v_spec_mean_da_cov = mean(v_spec_mean_da_cov),
