+++ vignettes/intro_CVtreeMLE.Rmd
@@ -128,7 +128,7 @@ niehs_results <- CVtreeMLE(data = as.data.frame(niehs_data),
                          w = c("Z", "Z2", "Z3"),
                          a = c(paste("X", seq(7), sep = "")),
                          y = "Y",
-                         fit_marginals = FALSE,
+                         fit_marginals = TRUE,
                          n_folds = 2,
                          seed = seed,
                          parallel_cv = FALSE,
