### 2022-10-19 ce11c9a
Including some print statements to monitor the simulations.
