+++ sandbox/03_run_2d_simulation_r1.R
@@ -82,6 +82,7 @@ for (sample_size in n_obs) {
   # get results in parallel
   results <- list()
   print(sample_size)
+
   for(this_iter in seq_len(n_sim)) {
     print(this_iter)
     stop <- FALSE
