+++ vignettes/intro_CVtreeMLE.Rmd
@@ -129,7 +129,7 @@ niehs_results <- CVtreeMLE(data = as.data.frame(niehs_data),
                          a = c(paste("X", seq(7), sep = "")),
                          y = "Y",
                          fit_marginals = TRUE,
-                         n_folds = 2,
+                         n_folds = 5,
                          seed = seed,
                          parallel_cv = FALSE,
                          family = "continuous",
