+++ .github/workflows/R-CMD-check.yml
@@ -2,9 +2,11 @@ on:
   push:
     branches:
       - main
+      - master
   pull_request:
     branches:
       - main
+      - master
 
 name: R-CMD-check
 
@@ -29,7 +31,7 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Setup R
-        uses: r-lib/actions/setup-r@main
+        uses: r-lib/actions/setup-r@master
         with:
           r-version: ${{ matrix.config.r }}
 
