+++ sandbox/04_make_plots.R
@@ -3,8 +3,9 @@ library(here)
 library(ggpubr)
 library(purrr)
 library(tidyr)
+library(hrbrthemes)
 
-n_obs <- c(200, 350, 500, 750, 1000, 1500, 2000, 3000)
+n_obs <- c(200, 350, 500, 750, 1000, 1500, 2000, 3000, 5000)
 
 sim_results_1 <- readRDS(
   here("sandbox/data/CVtreeMLE_run_1.rds")
@@ -22,33 +23,43 @@ sim_results_4 <- readRDS(
   here("sandbox/data/CVtreeMLE_run_4.rds")
 )
 
-sim_results <- rbind(sim_results_1, sim_results_2, sim_results_3, sim_results_4)
+sim_results_5 <- readRDS(
+  here("sandbox/data/CVtreeMLE_run_5.rds")
+)
+
+sim_results <- rbind(sim_results_2, sim_results_3)
 
 
 sim_statistics <- sim_results %>%
   group_by(n_obs) %>%
   summarise(
-    tmle_pooled_da_bias = abs(mean(tmle_pooled_da_bias)),
-    tmle_pooled_gt_bias = abs(mean(tmle_pooled_gt_bias)),
-    v_spec_da_mean_bias = abs(mean(v_spec_da_mean_bias)),
-    v_spec_gt_mean_bias = abs(mean(v_spec_gt_mean_bias)),
-    v_pooled_da_bias = abs(mean(v_pooled_da_bias)),
-    v_pooled_gt_bias = abs(mean(v_pooled_gt_bias)),
+    abs_tmle_pooled_da_bias = abs(mean(tmle_pooled_da_bias)),
+    abs_tmle_pooled_gt_bias = abs(mean(tmle_pooled_gt_bias)),
+    abs_v_spec_da_mean_bias = abs(mean(v_spec_da_mean_bias)),
+    abs_v_spec_gt_mean_bias = abs(mean(v_spec_gt_mean_bias)),
+    abs_v_pooled_da_bias = abs(mean(v_pooled_da_bias)),
+    abs_v_pooled_gt_bias = abs(mean(v_pooled_gt_bias)),
     tmle_pooled_sd = sd(tmle_pooled_ate),
     v_spec_mean_sd = sd(v_spec_mean_ate),
     v_pooled_mean_sd = sd(v_pooled_ate),
-    tmle_pooled_da_mse = tmle_pooled_da_bias^2 + tmle_pooled_sd^2,
-    tmle_pooled_gt_mse = tmle_pooled_gt_bias^2 + tmle_pooled_sd^2,
-    v_spec_da_mean_mse = v_spec_da_mean_bias^2 + v_spec_mean_sd^2,
-    v_spec_gt_mean_mse = v_spec_gt_mean_bias^2 + v_spec_mean_sd^2,
-    v_pooled_da_mse = v_pooled_da_bias^2 + v_pooled_mean_sd^2,
-    v_pooled_gt_mse = v_pooled_gt_bias^2 + v_pooled_mean_sd^2,
+    tmle_pooled_da_mse = abs_tmle_pooled_da_bias^2 + tmle_pooled_sd^2,
+    tmle_pooled_gt_mse = abs_tmle_pooled_gt_bias^2 + tmle_pooled_sd^2,
+    v_spec_da_mean_mse = abs_v_spec_da_mean_bias^2 + v_spec_mean_sd^2,
+    v_spec_gt_mean_mse = abs_v_spec_gt_mean_bias^2 + v_spec_mean_sd^2,
+    v_pooled_da_mse = abs_v_pooled_da_bias^2 + v_pooled_mean_sd^2,
[…]
