### 2022-11-03 0cf91a1
Update 3d mixture cube simulation to have mean 0 covariates for easier calculation of ground-truth. Wrote first 3d simulation script for testing.
