### 2022-11-03 f763833
3d sim slurm run script
