+++ sandbox/03_run_3d_simulation_r1.R
@@ -9,6 +9,7 @@ library(foreach)
 
 library(CVtreeMLE)
 source(here("sandbox", "02_fit_estimators.R"))
+source(here("sandbox", "simulate_2d_data.R"))
 
 Sys.unsetenv("GITHUB_PAT")
 devtools::install(upgrade = "never")
