### 2022-11-03 267fa7f
output update for 3d sim
