+++ sandbox/03_run_3d_simulation_r1.R
@@ -77,4 +77,4 @@ for (sample_size in n_obs) {
 # save results to file
 timestamp <- str_replace_all(Sys.time(), " ", "_")
 saveRDS(object = sim_results_df,
-        file = here("sandbox/data", paste0("CVtreeMLE_", "run_1", ".rds")))
+        file = here("sandbox/data", paste0("3D_CVtreeMLE_", "run_1", ".rds")))
