### 2022-11-04 #4181 @blind-contours
I’ve made some big updates to the package. Currently passes everything with no notes, warnings or errors but it appears something is failed the check with GitHub Linux which I’m trying to figure out. I think it’s just the parallel processing in the vignette on Linux. I think I’ve responded to everything else. Let me know if I can do anything else.
