### 2022-11-07 3cb3642
Split up the 3d sims for 5 runs of 10 iters each.
