+++ R/est_marg_nuisance_params.R
@@ -114,15 +114,10 @@ est_marg_nuisance_params <- function(at,
           folds = 2
         )
 
-        discrete_sl_metalrn <- sl3::Lrnr_cv_selector$new(
-          sl3::loss_loglik_binomial)
+        sl <- sl3::Lrnr_sl$new(learners = aw_stack,
+                               metalearner = sl3::Lrnr_nnls$new())
 
-        discrete_sl <- sl3::Lrnr_sl$new(
-          learners = aw_stack,
-          metalearner = discrete_sl_metalrn,
-        )
-
-        sl_fit <- suppressWarnings(discrete_sl$train(task_at))
+        sl_fit <- suppressWarnings(sl$train(task_at))
 
         ghat_1w <- bound_precision(sl_fit$predict(task_av))
 
@@ -168,15 +163,10 @@ est_marg_nuisance_params <- function(at,
           outcome_type = family
         )
 
-        discrete_sl_metalrn <- sl3::Lrnr_cv_selector$new(
-          sl3::loss_squared_error)
+        sl <- sl3::Lrnr_sl$new(learners = aw_stack,
+                               metalearner = sl3::Lrnr_nnls$new())
 
-        discrete_sl <- sl3::Lrnr_sl$new(
-          learners = aw_stack,
-          metalearner = discrete_sl_metalrn,
-        )
-
-        sl_fit <- suppressWarnings(discrete_sl$train(task_at))
+        sl_fit <- suppressWarnings(sl$train(task_at))
 
         q_bar_aw <- sl_fit$predict(task_av)
         q_bar_1w <- sl_fit$predict(task_av_1)
