+++ sandbox/NHANES/code/2_nhanes_var_procesing.R
@@ -1,0 +1,682 @@
+################################################################################
+### Title:        NHANES Variable Processing
+### Author:        Aidan
+### Date Created:  4/14/2020
+################################################################################
+
+
+
+################################################################################
+### I. Environment, data loading
+################################################################################
+
+`%p%` <- function(x,y) paste0(x,y)
+library(tidyverse)
+library(magrittr)
+library(janitor)
+library(readxl)
+library(purrr)
+library(gtools)
+library(here)
+
+### get variable list
+var_list <-
+  read_excel(here("sandbox/NHANES/input/Dictionary_1516.xlsx"), sheet = "selected",
+             col_names = FALSE)
+
+names(var_list) <-
+  c("var_name", "var_description")
+
+
+### get data dictionary
+data_dic <-
+  read_excel(here("sandbox/NHANES/input/Dictionary_1516.xlsx"), sheet = "total",
+             col_names = TRUE)
+
+
+names(data_dic) <-
+  c("var_name", "var_description")
+
+
+### get relevant NHANES data
+rds_files <-
+  list.files(here("sandbox/NHANES/input/"),pattern="data.rds")
+
+
+empty_df <- c()
+
+
+### manually adding some dietary variables for now.
+diet_vars <- c("DR1TCAFF", "DR2TCAFF", "DRXTCAFF",
+               "DR1TALCO", "DR2TACLO", "DRXTALCO",
+               "DR1TPROT", "DR2TPROT", "DRXTPROT",
+               "DR1TCHOL", "DR2TCHOL", "DRXTCHOL",
+               "DR1TSFAT", "DR2TSFAT", "DRXTSFAT",
+               "DR1TMFAT", "DR2TMFAT", "DRXTMFAT",
+               "DR1TPFAT", "DR2TPFAT", "DRXTPFAT")
+
+full_var_list <-
+  append(var_list$var_name, diet_vars)
+
+
+for(paths in rds_files) {
+
+  assign(gsub("_data.rds", "", paths),
+
+         readRDS(here("sandbox/NHANES/input/") %p% paths) %>%
+           dplyr::select(matches(paste(full_var_list, collapse = "|"))) %>%
+           mutate(name = gsub("_data.rds", "", paths))
+  )
+
+  if(ncol(get(gsub("_data.rds", "", paths))) <= 2) {
+    empty_df <-
+      append(empty_df,
+             gsub("_data.rds", "", paths))}
+
+
+}
+
+### remove data if no variables of interest
+rm(list=empty_df)
+
+### some variables appear twice in the Glucose datasets.
+for(g in c("B", "C", "D", "E", "F", "G", "H", "I")) {
+  if("GLU_" %p% g %in% ls()) {
+    assign("GLU_" %p% g,
+           dplyr::select(get("GLU_" %p% g),
+                  -contains("PHAFST")))
+  }
+}
+
+### redundant tables
+rm(COT_E)
+
+
+################################################################################
+### II. Data Merging
+################################################################################
+
+
+### Get list of data
+data_list = list()
+names <- list()
+j=1
+for(i in ls()) {
+
+  if("data.frame" %in% class(get(i)) & !i %in% c("var_list", "data_dic")) {
+    data_list[[j]] = get(i) %>% mutate(name = i)
+    names[[j]] <- i
+    j=j+1
+  }
+}
+
+names(data_list) <- names
+
+data_list <-
+  lapply(data_list,
+       function(x) x %>%
+         dplyr::mutate(year_begin = dplyr::case_when(
+           grepl("_A", name) ~ 1999,
+           grepl("_B", name) ~ 2001,
+           grepl("_C", name) ~ 2003,
+           grepl("_D", name) ~ 2005,
+           grepl("_E", name) ~ 2007,
+           grepl("_F", name) ~ 2009,
[…]
