+++ R/est_mix_nuisance_params_nde.R
@@ -1,0 +1,316 @@
+#' @title Estimate the NDE nuisance parameters for each mixture
+#'  interaction identified
+#'
+#' @description For each mixture mixture interaction found, create a g estimator
+#' for the probability of being exposed to the rule thresholds,
+#'a Q estimator for the outcome E(Y| A = a_mix, Z, W), an e estimator for
+#' e(A|Z,W) and a Phi estimator that regresses the outcome counterfactuals on
+#' the covariates of A =0.  Get estimates of g, e,phi, and Q using the
+#' validation data and calculate the clever covariate used in the
+#' TMLE fluctuation step.
+#'
+#' @param at Training data
+#' @param av Validation data
+#' @param w Vector of characters denoting covariates
+#' @param z Vector of characters denoting mediators
+#' @param y The outcome variable
+#' @param no_mix_rules TRUE/FALSE indicator for if no mixture rules were found
+#'
+#' @param aw_stack Super Learner library for fitting Q (outcome mechanism) and
+#' g (treatment mechanism)
+#' @param psi_z_stack Super Learner library for fitting Q_diff (difference
+#' in conditional outcomes for A=1 and A = 0, and Z = z, W = w) onto W of
+#' controls observations
+#' @param family Binomial or continuous
+#' @param rules Dataframe of rules found during the PRE fitting process
+#' @param h_aw_trunc_lvl Truncation level of the clever covariate (induces more
+#'  bias to reduce variance)
+#' @param parallel_cv TRUE/FALSE if cv parallelization is used
+#' @param seed Seed number
+#' @importFrom magrittr %>%
+#' @importFrom dplyr group_by filter top_n
+#' @return A list of dataframes where the nuisance parameters are added to
+#' the raw data.
+
+#'
+#' @export
+
+est_mix_nuisance_params_nde <- function(at,
+                                        av,
+                                        w,
+                                        z,
+                                        y,
+                                        no_mix_rules,
+                                        aw_stack,
+                                        psi_z_stack,
+                                        family,
+                                        rules,
+                                        h_aw_trunc_lvl = 10,
+                                        parallel_cv,
+                                        seed) {
+  if (parallel_cv == TRUE) {
+    future::plan(future::sequential, gc = TRUE)
+  }
+
+  set.seed(seed)
+
+  at_mix <- at
+  av_mix <- av
+
+  if (no_mix_rules != TRUE) {
+    at_rules_eval <-
+      evaluate_mixture_rules(data = at_mix, rules = rules)
+    av_rules_eval <-
+      evaluate_mixture_rules(data = av_mix, rules = rules)
+
+    mix_interaction_data <- list()
+
+    for (interaction in seq(dim(at_rules_eval)[2])) {
+      interaction_rule <- at_rules_eval[, interaction]
+
+      if (dim(table(interaction_rule)) == 2) {
+
+        at_mix$A_mix <- interaction_rule
+        av_mix$A_mix <- av_rules_eval[, interaction]
+
+        # get g mech estimates ---------------------------
+
+        task_at_g <- sl3::make_sl3_Task(
+          data = at_mix,
+          covariates = w,
+          outcome = "A_mix",
+          outcome_type = "binomial",
+          folds = 2
+        )
+
[…]
