+++ vignettes/intro_CVtreeMLE.Rmd
@@ -128,8 +128,8 @@ niehs_results <- CVtreeMLE(data = as.data.frame(niehs_data),
                          w = c("Z", "Z2", "Z3"),
                          a = c(paste("X", seq(7), sep = "")),
                          y = "Y",
-                         fit_marginals = FALSE,
-                         n_folds = 10,
+                         fit_marginals = TRUE,
+                         n_folds = 5,
                          seed = seed,
                          parallel_cv = TRUE,
                          parallel = TRUE,
