+++ NAMESPACE
@@ -40,6 +40,8 @@ export(simulate_mixture_cube)
 export(v_fold_marginal_qgroup_split)
 export(v_fold_mixture_group_split)
 import(furrr)
+import(hrbrthemes)
+import(viridis)
 importFrom(MASS,mvrnorm)
 importFrom(assertthat,assert_that)
 importFrom(cvTools,cvFolds)
@@ -69,4 +71,5 @@ importFrom(stats,qnorm)
 importFrom(stats,qunif)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
+importFrom(stats,sd)
 importFrom(stringr,str_detect)
