### 2022-11-24 8fe45e9
Updating clever covariate thresholds and code formatting.
