+++ R/est_mix_nuisance_params.R
@@ -36,7 +36,7 @@ est_mix_nuisance_params <- function(at,
                                     aw_stack,
                                     family,
                                     rules,
-                                    h_aw_trunc_lvl = 10,
+                                    h_aw_trunc_lvl = h_aw_trunc_lvl,
                                     parallel_cv,
                                     seed) {
   if (parallel_cv == TRUE) {
@@ -78,17 +78,21 @@ est_mix_nuisance_params <- function(at,
           outcome_type = "binomial"
         )
 
-        sl <- sl3::Lrnr_sl$new(learners = aw_stack,
-                               metalearner = sl3::Lrnr_nnls$new())
+        sl <- sl3::Lrnr_sl$new(
+          learners = aw_stack,
+          metalearner = sl3::Lrnr_nnls$new()
+        )
 
         sl_fit <- suppressWarnings(sl$train(task_at))
 
         ghat_1w <- sl_fit$predict(task_av)
 
-        h_aw <- calc_clever_covariate(ghat_1_w = ghat_1w,
-                                      data = av_mix,
-                                      exposure = "A_mix",
-                                      h_aw_trunc_lvl = 10)
+        h_aw <- calc_clever_covariate(
+          ghat_1_w = ghat_1w,
+          data = av_mix,
+          exposure = "A_mix",
+          h_aw_trunc_lvl = h_aw_trunc_lvl
+        )
 
         task_at <- sl3::make_sl3_Task(
           data = at_mix,
@@ -123,8 +127,10 @@ est_mix_nuisance_params <- function(at,
           outcome_type = family
         )
 
-        sl <- sl3::Lrnr_sl$new(learners = aw_stack,
-                               metalearner = sl3::Lrnr_nnls$new())
+        sl <- sl3::Lrnr_sl$new(
+          learners = aw_stack,
+          metalearner = sl3::Lrnr_nnls$new()
+        )
 
         sl_fit <- suppressWarnings(sl$train(task_at))
 
