### 2022-11-25 a86ab71
udpate truncation
