### 2022-12-13 #4181 @cpalmer718
@osorensen There are still no contributing guidelines and the corresponding link in the readme is broken. I'm fine with the paper at this point, effort has obviously been put in here.
