### 2022-12-13 55db18f
Create contributing.md
Adding contributing guidelines to help users understand how to contribute to CVtreeMLE!
