### 2022-12-13 #4181 @osorensen
Thanks a lot @cpalmer718.

@blind-contours, could you please fix the two remaining issues pointed out by @cpalmer718 and report here when done?
