### 2023-01-02 #4181 @osorensen
> Thanks a lot @cpalmer718.
>
> @blind-contours, could you please fix the two remaining issues pointed out by @cpalmer718 and report here when done?

@blind-contours, could you please update us on this one?

@GaryBAYLOR could you please update us on how it's going with your review?
