### 2023-01-04 d1f411f
Updating contributing, vignette, update tests.
