+++ man/est_marg_nuisance_params.Rd
@@ -14,9 +14,9 @@ est_marg_nuisance_params(
   a,
   no_marg_rules,
   marg_decisions,
-  h_aw_trunc_lvl = 10,
   parallel_cv,
-  seed
+  seed,
+  h_aw_trunc_lvl
 )
 }
 \arguments{
@@ -41,12 +41,12 @@ found across the folds}
 \item{marg_decisions}{List of rules found within the fold for each
 mixture component}
 
-\item{h_aw_trunc_lvl}{Truncation level of the clever covariate
-(induces more bias to reduce variance)}
-
 \item{parallel_cv}{TRUE/FALSE if cv parallelization is used}
 
 \item{seed}{Seed number}
+
+\item{h_aw_trunc_lvl}{Truncation level of the clever covariate
+(induces more bias to reduce variance)}
 }
 \value{
 marginal_data: A list of data frames for each mixture component
