### 2023-01-04 #4181 @blind-contours
@osorensen I have added contributing guidelines and the corresponding link now maps to my website. Package should pass with no notes, warnings, or errors. I'm still making small changes to the vignette and general code streamlining but I think I've responded to all the comments/suggestions at this point. Thank you!
