+++ R/CVtreeMLE.R
@@ -158,15 +158,17 @@
 #' colnames(x) <- c("A1", "A2", "W1", "W2")
 #' y_prob <- plogis(3 * sin(x[, 1]) + sin(x[, 2]), sin(x[, 4]))
 #' Y <- rbinom(n = n, size = 1, prob = y_prob)
-#' data <- as.data.frame(cbind(x,Y))
+#' data <- as.data.frame(cbind(x, Y))
 #'
-#' CVtreeMLE_fit <- CVtreeMLE(data = data,
-#'                           w = c("W1", "W2"),
-#'                           a = c("A1", "A2"),
-#'                           y = "Y",
-#'                           family = "binomial",
-#'                           parallel = FALSE,
-#'                           n_folds = 2)
+#' CVtreeMLE_fit <- CVtreeMLE(
+#'   data = data,
+#'   w = c("W1", "W2"),
+#'   a = c("A1", "A2"),
+#'   y = "Y",
+#'   family = "binomial",
+#'   parallel = FALSE,
+#'   n_folds = 2
+#' )
 #'
 #' @return Object of class \code{CVtreeMLE}, containing a list of table results
 #' for: marginal ATEs, mixture ATEs, RMSE of marginal model fits, RMSE of
@@ -216,7 +218,6 @@ CVtreeMLE <- function(w,
                       max_iter = 5,
                       verbose = FALSE,
                       h_aw_trunc_lvl = 10) {
-
   if (any(sapply(data[, a], is.factor))) {
     print("Factor variable detected in exposures, converting to numeric")
     data[, a] <- sapply(data[, a], as.numeric)
@@ -278,35 +279,33 @@ CVtreeMLE <- function(w,
   data$folds <- create_cv_folds(n_folds, data[, y])
 
   if (parallel == TRUE) {
-    if(parallel_type == "multi_session") {
-    future::plan(future::multisession,
-                 workers = num_cores,
-                 gc = TRUE)
-    }else {
+    if (parallel_type == "multi_session") {
+      future::plan(future::multisession,
+        workers = num_cores,
+        gc = TRUE
+      )
+    } else {
       future::plan(future::multicore,
-                   workers = num_cores,
-                   gc = TRUE)
+        workers = num_cores,
+        gc = TRUE
+      )
     }
   } else {
     future::plan(future::sequential,
-                 gc = TRUE)
+      gc = TRUE
+    )
   }
 
   data <- as.data.frame(data)
 
+  # Set up lists to store data in the CV
+
   fold_results_mix_rules <- list()
-  fold_results_marg_rules <- list()
-  fold_results_marg_directions <- list()
-
   mix_fold_data <- list()
-  mix_fold_directions <- list()
 
+  fold_results_marg_rules <- list()
   fold_results_marginal_data <- list()
 
-  fold_results_marginal_combo_data <- list()
-
-  fold_results_mix_combo_data <- list()
-  fold_results_mix_Sls <- list()
 
   # Iterative Back-fitting on Mixture ---------------------------
 
@@ -315,42 +314,50 @@ CVtreeMLE <- function(w,
   }
 
   fold_mixture_rules <- furrr::future_map(unique(data$folds),
-                                              function(fold_k) {
+    function(fold_k) {
+      at <- data[data$folds != fold_k, ]
 
-    at <- data[data$folds != fold_k, ]
+      rules <-
+        fit_mix_rule_backfitting(
+          at = at,
+          a = a,
+          w = w,
+          y = y,
+          direction = direction,
+          w_stack,
+          fold = fold_k,
+          max_iter,
+          verbose,
+          parallel_cv = parallel_cv,
+          seed = seed
+        )
 
-    rules <-
-      fit_mix_rule_backfitting(
-        at = at,
-        a = a,
[…]
