### 2023-01-04 #4181 @osorensen
👋 @SaranjeetKaur @calebsjin, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

The review of this submission has been going on for a while and the authors has addressed a number of issues, but we have lost contact with one of the reviewers. We would hence very much appreciate the opinion of one or two additional reviewers.
