### 2023-01-05 #4181 @cpalmer718
@osorensen I see the markdown file in the repo, which is good, but the links to it in the readme are broken: e.g. should be `https://github.com/blind-contours/CVtreeMLE/blob/main/contributing.md` but instead is `https://github.com/blind-contours/%20CVtreeMLE/blob/main/CONTRIBUTING.md` (note floating space). Additionally, the article citation in the markdown readme seems to be malformed.

Regardless, I see the contributing doc, so while the above should be fixed, I'll check the thing off. I'm sorry this has languished.
