### 2023-01-05 #4181 @blind-contours
Thanks @cpalmer718 I have fixed the issue with the contributing link and have simply removed the citation for now until we get our paper on Arxiv which should happen soon. I will update the readme with the JOSS citation and Arxiv citation once complete.
