### 2023-01-16 blind-contours/CVtreeMLE#23 @wleoncio
Long installation time
Not necessarily a deal-breaker for using this package, but an annoyance especially when one needs to install the package multiple times (for testing on a machine or for use on multiple ones): first installation on my 8th-gen i5 CPU took no less than 30 minutes.

I'm glad to see a [CRAN release on the roadmap](https://github.com/openjournals/joss-reviews/issues/4181#issuecomment-1227593685), hopefully this will reduce the amount of dependencies and compiled code involved.

Is there anything more short-term that could be done, though?
