### 2023-01-16 blind-contours/CVtreeMLE#24 @wleoncio
Missing dependencies: improve installation instructions?
# Installation dependencies

Installation was straightforward from my development machine, but starting from a clean R session (using a [Rocker image](https://rocker-project.org/)), running `remotes::install_github("blind-contours/CVtreeMLE@main")` took me down a rabbit hole of dependency-chasing which I just gave up on after a couple of days trying.

It's hard to say how many users would struggle with installing this package, but I would not assume they will have the full-blown machines software developers use. Given how long the installation takes (issue #23), I would try to make sure the user has all the dependencies set up before trying to install CVtreeMLE, or at the very least warn them. Either that, or explicitly restrict the OSs your package supports (desperate measure, not really recommended).

Conversely, since your CI checks are testing for Ubuntu, I guess installation on a clean Linux machine is still fine. I do, however, see some [hard-coded dependency handling there](https://github.com/blind-contours/CVtreeMLE/actions/runs/3849846438/workflow#L47-L50) that might hint about missing dependencies in the README file.

# Example execution dependencies

Even from my development PC, I couldn't run the example straight off because I had to install the following packages:

- `kableExtra`
- `ranger`
- `xgboost`
- `nnls` (this one was particularly annoying since it made the `CVtreeMLE()` call fail only after almost 20 minutes of computing time)

I wish these extra dependencies would also have been mentioned somewhere (`kableExtra` is a suggested package, but still should be explicitly mentioned for running examples).
