### 2023-01-16 blind-contours/CVtreeMLE#25 @wleoncio
Add Windows to build CI workflow
I noticed the [build CI file](https://github.com/blind-contours/CVtreeMLE/actions/runs/3849846438/workflow) contains Linux and macOS builds. It would be nice to also include a Windows check in there (even if that's the platform you're developing on), so that the tests are more comprehensibe by including non-Unix OSs, and would benefit contributors not on Windows machines (there are dozens of us. Dozens!)
