### 2023-01-16 blind-contours/CVtreeMLE#26 @wleoncio
Reduce number of dependencies
Eyeing a potential CRAN submission (or at least a welcome bloat-reduction), would you consider reducing the number of dependencies in your package 20+ is _a lot_, IMO)?
