### 2023-01-16 blind-contours/CVtreeMLE#27 @wleoncio
Simplify unit tests
Running `devtools::test()` required the installation of `mlbench` package. Also, I had to manually tweak the tests to use more CPUs to bring computing time to an acceptable state.

This is something that would only bother developers, possibly scaring away contributors, so not a big deal, but it would be nice for unit tests to be faster (perhaps by using smaller datasets, fewer iterations, more CPUs) and more complex (most of them just check output class or triggered errors).
