+++ .github/workflows/R-CMD-check.yml
@@ -21,6 +21,7 @@ jobs:
         config:
           - {os: ubuntu-latest, r: 'release'}
           - {os: macOS-latest, r: 'release'}
+          - {os: windows-latest, r: 'release'}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
