### 2023-01-25 blind-contours/CVtreeMLE#24 @blind-contours
Hi @wleoncio - thanks for catching this and reviewing the package in general. To answer your comment here, I've added a small section in the readme that tells the user to install the dependencies that are hard coded in the YML file you mentioned. Sorry I forgot about this, SL3 has those ranger and nnls dependencies that for some reason need to be installed before SL3 and likewise, because CVtreeMLE comes with plotting features, there are dependencies for that. I've also added all these packages to the suggests section in the DESCRIPTION file. Let me know if you'd like anything else done on this front.
