### 2023-01-25 blind-contours/CVtreeMLE#25 @blind-contours
Hi @wleoncio - I've added windows to the build file. I think originally I removed this because a version of SL3@devel was failing install on windows and I couldn't figure out why (not related to CVtreeMLE). I've added it back and I'll see if it passes the tests on Windows now.
