### 2023-01-25 blind-contours/CVtreeMLE#23 @blind-contours
Hi @wleoncio - I found 2 packages that are used in the tests and can be moved to suggests so I did that. I think the other thing I can do to remove the number of dependencies is to make the simulation scripts used in the readme, vignette, and paper not actually part of the package which would cut back 1 package in imports. I can try and think of some other things to try and reduce install time. The issue is that the CVtreeMLE package relies on a lot of other packages that take time to install namely, PRE, partykit, SL3, furrr etc. Let me know if you have some other ideas.
