### 2023-01-25 blind-contours/CVtreeMLE#26 @blind-contours
This package requires a lot of dependencies by it's nature - integrating trees as a data-adaptive parameter (pre, partykit etc.) with machine learning (sl3, ranger, and others) then all the data formatting and parallelization (furrr, data.table, purr, stringr, stats, dplyr etc.). I see a couple ways moving forward where I can clean some things up. 1. Move simulations out of the package, this knocks out 1-2 packages that are required and moves them to suggests, 2. Getting rid of the plotting functions that come default in the package would get rid of 3. Also making the readme less pretty with knit I could also get rid of that. I'll make these changes before the submission to CRAN. Also SL3 is being updated for CRAN submission and that should get rid of the misc. dependencies we need to install for that package (should be packaged in SL3).
