+++ sandbox/NHANES/code/NHANES_example.R
@@ -1,6 +1,7 @@
 library(stringr)
 library(zoo)
 library(imputeTS)
+library(here)
 
 nhanes_data <- readRDS(here("sandbox/NHANES/output/NHANES_data.RDS"))
 
