### 2023-01-30 blind-contours/CVtreeMLE#25 @wleoncio
Thanks for taking care of this. The Windows check seems to still contain an SL3-related issue, but I would rather have an issue opened to take care of it later than drop the OS check altogether.
