### 2023-01-30 blind-contours/CVtreeMLE#27 @wleoncio
Tests are still taking an inordinate amount of time on my machine (8th gen i5 CPU), see screenshot below:

![20min](https://user-images.githubusercontent.com/8234768/215482007-5e702b97-72dc-46fc-8e45-bcaa46807fe2.png)

How much time does the whole `devtools::test()` procedure take on your side? Since the breast cancer test takes half as long as I got, I'm guessing the whole thing takes around 10 minutes.
