### 2023-01-30 blind-contours/CVtreeMLE#23 @wleoncio
Thank you for addressing this issue. I understand there's not much that can be done as long as there are so many dependencies (see issue #26), so I guess a long installation is something users will have to live with for a while.
