### 2023-01-30 #4181 @wleoncio
Hi @osorensen,

I think we've made some progress, issues [24](https://github.com/blind-contours/CVtreeMLE/issues/24) and [25](https://github.com/blind-contours/CVtreeMLE/issues/25) are taken care of, and [23](https://github.com/blind-contours/CVtreeMLE/issues/23) and [26](https://github.com/blind-contours/CVtreeMLE/issues/26) may be left as future improvements.

Regarding [27](https://github.com/blind-contours/CVtreeMLE/issues/27), however, I'd like to hear back from @blind-contours once more about the possibility of further reducing test times.
