### 2023-01-30 blind-contours/CVtreeMLE#27 @blind-contours
Hi @wleoncio - attached is my run:

![Screen Shot 2023-01-30 at 9 44 04 AM](https://user-images.githubusercontent.com/54640903/215553512-5286b96d-1099-4469-a4d9-c1a23ce6307c.png)
