### 2023-01-30 blind-contours/CVtreeMLE#27 @blind-contours
So it's about 12 minutes - is that reasonable? I can exchange some things if that feels better.
