### 2023-01-31 blind-contours/CVtreeMLE#27 @wleoncio
Thanks for the extra info, @blind-contours! It would be nice those those mixture tests closer to the minute mark, but in the end it's up to you how much time you want to spend optimizing this part of the code that doesn't directly affect users.
