### 2023-02-11 blind-contours/CVtreeMLE#18 @GaryBAYLOR
I checked the above issue still exists in `calc_v_fold_mixtures_ate.Rd` which has the following lines
```r
\value{
Rules object. TODO: add more detail here.
```
The reason is that in the file `calc_v_fold_mixture_ate.R` we can see there is a duplicate return. Removing the first return and regenerate Rd file will likely solve the problem.
```r
#' @return Rules object. TODO: add more detail here.
#' @importFrom stats as.formula glm p.adjust plogis predict qlogis qnorm
#' @importFrom stats qunif rnorm runif
#' @importFrom rlang :=
#' @return A data frame of mixture results for each variable level combination
#' for each fold
```
