### 2023-02-11 ba2245f
Change based on JOSS issue 18
