### 2023-02-12 blind-contours/CVtreeMLE#2 @GaryBAYLOR
The license in the DESCRIPTION file is still: `MIT + file LICENSE`, not aligning with the one in LICENSE file and README, which is MIT only
