### 2023-02-12 #4181 @GaryBAYLOR
Hi @osorensen, thank you for your patience in waiting for my review results. I checked all previous open issues and added a new one today, all of which have now been addressed by the author with the exception of only this one:  https://github.com/blind-contours/CVtreeMLE/issues/2.After this is addressed, I am good on my end.

Refer to https://github.com/blind-contours/CVtreeMLE/issues?q=is%3Aissue+is%3Aclosed for all closed issues.
