### 2023-02-12 #4181 @blind-contours
I just addressed this issue related to the license: https://github.com/blind-contours/CVtreeMLE/issues/2
