### 2023-02-16 blind-contours/CVtreeMLE#29 @osorensen
Editorial comment on JOSS submission
This is an editorial comment on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4181). I'll post my comments as separate issues while reading through the manuscript. This is the first one.

One lines 17-18, you write

> more flexible methods [@Bobb2014] lack statistical inference

Could you please clarify what you mean by this? I briefly looked at the paper by Bobb et al., and it seems to me that their method yields posterior distributions which can be used for statistical inference. Are you thinking particularly about frequentist inference, or are there other issues which I'm not catching here?
