### 2023-02-16 blind-contours/CVtreeMLE#30 @osorensen
Is this sentence missing something?
Another editorial comment on your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4181).

On lines 84-87 you write

> In the case of mixtures, it is necessary to map a set of continuous mixture components into a lower dimensional representation of exposure using a pre-determined algorithm then estimate a target parameter on this more interpretable exposure.

This is hard to read. I suggest adding ", and" between "algorithm" and "then", so it becomes

> In the case of mixtures, it is necessary to map a set of continuous mixture components into a lower dimensional representation of exposure using a pre-determined algorithm, and then estimate a target parameter on this more interpretable exposure.
