### 2023-02-16 blind-contours/CVtreeMLE#31 @osorensen
Reference to Zheng and van der Laan paper
This is an editorial comment on your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4181).

When clicking the [link](https://biostats.bepress.com/ucbbiostat/paper273/) to the Zheng and van der Laan reference, I find this note:

> Comments
This material is published in: W. Zheng, M.J. van der Laan (2011). "Cross-Validated Targeted Minimum-Loss-Based Estimation." In M.J. van der Laan and S. Rose, Targeted Learning: Causal Inference for Observational and Experimental Data, Chapter 27. New York, Springer. This work is supported by NIH Targeted Empirical Super Learning in AIDS & Epidemiology grant # 5R01AI74345-5

Could you please update the citation so it cites this published book chapter rather than the preprint?
