### 2023-02-16 blind-contours/CVtreeMLE#32 @osorensen
Editorial comment on JOSS submission
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4181).

On lines 75-76 you write

> In most research scenarios, the analyst is interested in causal inference for an a priori specified treatment or exposure.

This seems to me a pretty strong statement. Could you perhaps write "In **many** research scenarios ..." instead?
