### 2023-02-16 blind-contours/CVtreeMLE#33 @osorensen
Editorial comment on JOSS submission
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4181):

You write in Statement of Need

> Current software tools for mixtures rarely report performance tests using data that reflect the complexities of real-world exposures.

Could you please provide references to these existing tools?
