### 2023-02-16 blind-contours/CVtreeMLE#29 @blind-contours
Thanks, this is a good point - I wasn't clear here. I have changed this sentence to:

More flexible methods such as Bayesian kernel machine regression (BKMR)[@Bobb2014] are sensitive to the choice of tuning parameters, are computationally taxing and lack an interpretable and robust summary statistic of dose-response relationships.
