### 2023-02-16 blind-contours/CVtreeMLE#33 @blind-contours
I added some citations here, specifically for the quantile g-comp method which does not provide adequate simulation data that reflect real exposure situations. I've also added a review article of methods that doesn't directly specify this problem but does discuss limitations of many methods, and I've also added an NIEHS article that discusses the needed for better vetted methods against some ground truth.
