+++ inst/REFERENCES.bib
@@ -1,3 +1,47 @@
+@article{carlin2013unraveling,
+  title={Unraveling the health effects of environmental mixtures: an NIEHS priority},
+  author={Carlin, Danielle J and Rider, Cynthia V and Woychik, Richard and Birnbaum, Linda S},
+  journal={Environmental health perspectives},
+  volume={121},
+  number={1},
+  pages={A6--A8},
+  year={2013},
+  publisher={NIH Public Access},
+  doi={10.1289/ehp.1206187}
+}
+
+@article{Keil2019,
+abstract = {Background: Exposure mixtures frequently occur in data across many domains, particularly in the fields of environmental and nutritional epidemiology. Various strategies have arisen to answer questions about exposure mixtures, including methods such as weighted quantile sum (WQS) regression that estimate a joint effect of the mixture components. Objectives: We demonstrate a new approach to estimating the joint effects of a mixture: quantile g-computation. This approach combines the inferential simplicity of WQS regression with the flexibility of g-computation, a method of causal effect estimation. We use simulations to examine whether quantile g-computation and WQS regression can accurately and precisely estimate effects of mixtures in a variety of common scenarios. Methods: We examine the bias, confidence interval coverage, and bias-variance tradeoff of quantile g-computation and WQS regression, and how these quantities are impacted by the presence of non-causal exposures, exposure correlation, unmeasured confounding, and non-linearity of exposure effects. Results: Quantile g-computation, unlike WQS regression allows inference on mixture effects that is unbiased with appropriate confidence interval coverage at sample sizes typically encountered in epidemiologic studies, and when the assumptions of WQS regression are not met. Further, WQS regression can magnify bias from unmeasured confounding that might occur if important components of the mixture are omitted from analysis. Discussion: Unlike inferential approaches that examine effects of individual exposures, while holding other exposures constant, methods like quantile g-computation that can estimate the effect of a mixture are essential for understanding effects of potential public health actions that act on exposure sources. Our approach may serve to help bridge gaps between epidemiologic analysis and interventions such as regulations on industrial emissions or mining processes, dietary changes, or consumer behavioral changes that act on multiple exposures simultaneously.},
+archivePrefix = {arXiv},
+arxivId = {1902.04200},
+author = {Keil, Alexander P. and Buckley, Jessie P. and O'Brien, Katie M. and Ferguson, Kelly K. and Zhao, Shanshan and White, Alexandra J.},
+doi = {10.1097/01.ee9.0000606120.58494.9d},
+eprint = {1902.04200},
+file = {:Users/davidmccoy/Library/Application Support/Mendeley Desktop/Downloaded/Keil et al. - 2019 - A quantile-based g-computation approach to addressing the effects of exposure mixtures.pdf:pdf},
+issn = {23318422},
+journal = {arXiv},
+number = {April},
+pages = {1--10},
+title = {{A quantile-based g-computation approach to addressing the effects of exposure mixtures}},
+volume = {128},
+year = {2019}
+}
+
+@article{Yu2022,
[…]
