+++ paper/paper.bib
@@ -1,3 +1,30 @@
+@article{carlin2013unraveling,
+  title={Unraveling the health effects of environmental mixtures: an NIEHS priority},
+  author={Carlin, Danielle J and Rider, Cynthia V and Woychik, Richard and Birnbaum, Linda S},
+  journal={Environmental health perspectives},
+  volume={121},
+  number={1},
+  pages={A6--A8},
+  year={2013},
+  publisher={NIH Public Access},
+  doi={10.1289/ehp.1206187}
+}
+
+@article{Yu2022,
+abstract = {Environmental risk factors have been implicated in adverse health effects. Previous epidemiological studies on environmental risk factors mainly analyzed the impact of single pollutant exposure on health, while in fact, humans are constantly exposed to a complex mixture consisted of multiple pollutants/chemicals. In recent years, environmental epidemiologists have sought to assess adverse health effects of exposure to multi-pollutant mixtures based on the diversity of real-world environmental pollutants. However, the statistical challenges are considerable, for instance, multicollinearity and interaction among components of the mixture complicate the statistical analysis. There is currently no consensus on appropriate statistical methods. Here we summarized the practical statistical methods used in environmental epidemiology to estimate health effects of exposure to multi-pollutant mixture, such as Bayesian kernel machine regression (BKMR), weighted quantile sum (WQS) regressions, shrinkage methods (least absolute shrinkage and selection operator, elastic network model, adaptive elastic-net model, and principal component analysis), environment-wide association study (EWAS), etc. We sought to review these statistical methods and determine the application conditions, strengths, weaknesses, and result interpretability of each method, providing crucial insight and assistance for addressing epidemiological statistical issues regarding health effects from multi-pollutant mixture.},
+author = {Yu, Linling and Liu, Wei and Wang, Xing and Ye, Zi and Tan, Qiyou and Qiu, Weihong and Nie, Xiuquan and Li, Minjing and Wang, Bin and Chen, Weihong},
+doi = {10.1016/j.envpol.2022.119356},
+file = {:Users/davidmccoy/Downloads/1-s2.0-S026974912200570X-main.pdf:pdf},
+issn = {18736424},
+journal = {Environmental Pollution},
+keywords = {Bayesian kernel machine regression,Environmental epidemiology,Health effect,Multi-pollutant mixture,Statistical method},
+number = {January},
+pmid = {35487468},
+title = {{A review of practical statistical methods used in epidemiological studies to estimate the health effects of multi-pollutant mixture}},
+volume = {306},
+year = {2022}
+}
+
 @article{Bobb2014,
 abstract = {Because humans are invariably exposed to complex chemical mixtures, estimating the health effects of multi-pollutant exposures is of critical concern in environmental epidemiology, and to regulatory agencies such as the U.S. Environmental Protection Agency. However, most health effects studies focus on single agents or consider simple two-way interaction models, in part because we lack the statistical methodology to more realistically capture the complexity of mixed exposures. We introduce Bayesian kernel machine regression (BKMR) as a new approach to study mixtures, in which the health outcome is regressed on a flexible function of the mixture (e.g. air pollution or toxic waste) components that is specified using a kernel function. In high-dimensional settings, a novel hierarchical variable selection approach is incorporated to identify important mixture components and account for the correlated structure of the mixture. Simulation studies demonstrate the success of BKMR in estimating the exposure-response function and in identifying the individual components of the mixture responsible for health effects. We demonstrate the features of the method through epidemiology and toxicology applications.},
 author = {Bobb, Jennifer F. and Valeri, Linda and {Claus Henn}, Birgit and Christiani, David C. and Wright, Robert O. and Mazumdar, Maitreyi and Godleski, John J. and Coull, Brent A.},
