### 2023-02-16 #4181 @blind-contours
New PDF is compiled and in the package with these edits.
