### 2023-02-17 #4181 @blind-contours
The only last thing is getting the package to install on Windows via github actions, I keep getting this error:

cannot open file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpoZc4PS\xshell1a402bd4033': Permission denied

I can't find a solution to this - have any of the reviewers seen this Windows error before? I'd just like this to show green checks on all operating systems per reviewers requests.
