### 2023-02-17 #4181 @wleoncio
Doesn't ring a bell, but notice how the installation of TinyTeX also doesn't go smoothly.

Perhaps updating your r-lib setup versions (see template [here](https://github.com/r-lib/actions/tree/v2-branch/examples#standard-ci-workflow)) helps. The new versions may even allow you to remove some manual installation commands from your YAML as well (namely, TinyTeX and the package dependencies).
