+++ paper/paper.bib
@@ -40,18 +40,14 @@ title = {{Bayesian kernel machine regression for estimating the health effects o
 volume = {16},
 year = {2014}
 }
-@article{keil2020,
-author = {{Alexander P. Keil, Jessie P. Buckley, Katie M. O'Brien, Kelly K. Ferguson, Shanshan Zhao}, and Alexandra J. White},
-doi = {10.1289/EHP8739},
-file = {:Users/davidmccoy/Downloads/EHP5838.pdf:pdf},
-issn = {15529924},
-journal = {Environmental Health Perspectives},
-number = {3},
-pages = {1--10},
-pmid = {33688746},
-title = {{A quantile-based g-computation approach to addressing the effects of exposure mixtures}},
-volume = {129},
-year = {2021}
+@article{keil2020quantile,
+  title={A quantile-based g-computation approach to addressing the effects of exposure mixtures},
+  author={Keil, Alexander P and Buckley, Jessie P and O’Brien, Katie M and Ferguson, Kelly K and Zhao, Shanshan and White, Alexandra J},
+  journal={Environmental health perspectives},
+  volume={128},
+  number={4},
+  pages={047004},
+  year={2020}
 }
 
 @article{Athey2016,
