### 2023-02-17 #4181 @blind-contours
@osorensen

I've created a new release at https://github.com/blind-contours/CVtreeMLE/releases/tag/v1.0-joss

and a Zenodo archive of that release at https://zenodo.org/record/7651354

The DOI for that release is 10.5281/zenodo.7651354
