### 2023-02-18 #4181 @blind-contours
@wleoncio, @cpalmer718 , @GaryBAYLOR , @osorensen - thank you all for the time you have spent reviewing and improving this package. This is my first piece of software that operationalizes a statistical method I created and I have just learned so much! Really appreciate the support.
