### 2023-02-21 #4181 @Kevin-Mattheus-Moerman
@blind-contours I am the AEiC for this track and here to help process the last steps. I've checked the archive and paper and have some minor comments that need your attention:

On the archive:
- [ ] The version tag on your repository is called `v1.0-joss`, while here and on ZENODO it is listed as `v1.0`. These need to match. Please check if you can either: rename the release on GitHub to match `v1.0` OR update the version here (I can do this) and on ZENODO to match `v1.0-joss`.
- [ ] The ZENODO listed license (currently listed as ` Creative Commons Attribution 4.0 International`) should match the software license which is the MIT license.

On the paper:
- [ ] Check/fix `heterogenous` -> `heterogeneous`
