### 2023-02-21 #4181 @blind-contours
Hi @Kevin-Mattheus-Moerman, I've changed the version on ZENODO to v1.0-joss and changed the license to MIT. I've also made and pushed the fix for heterogeneous in the paper and updated the affiliations. Let me know if you need anything else.
